api: split api.test.ts into themed test files
Break the single ~3800-line api.test.ts (one root describe('api')) into 15
focused *.test.ts files, each with its own root describe, grouped by theme /
class under test: metadata, enums, types, connection, prepared, results,
config, instance-cache, appender, scalar-functions, sql-utils, table-names,
variant, uuid, varint.

Shared helpers (sleep, withConnection, assertColumns, isVectorType,
getColumnVector, assertVectorValues, assertValues, bigints) move to
test/util/testHelpers.ts. The timezone setup that was a single beforeAll in the
old root describe becomes setDefaultTimezone(), called from a beforeAll in each
file so every test keeps the same fixed TIMESTAMPTZ offset.

A few tests are regrouped to their natural home rather than their original
location: the concurrent prepared-statement segfault test goes to prepared, and
the DuckDBGeometryType CRS test (previously nested in the VARIANT describe only
by position) goes to types. Redundant section-divider banner comments are
dropped; the DuckDBUUIDValue.fromUint128 doc comment travels with the uuid file.

Pure test reorganization: same 140 passing / 1 skipped, no test bodies changed.
