--- api/test/api.test.ts
@@ -1,3838 +1,0 @@
-import fs from 'fs';
-import { assert, beforeAll, describe, test } from 'vitest';
-import {
-  ANY,
-  ARRAY,
-  BIGINT,
-  BIGNUM,
-  BIT,
-  BLOB,
-  BOOLEAN,
-  DATE,
-  DECIMAL,
-  DOUBLE,
-  DateParts,
-  DuckDBArrayVector,
-  DuckDBBigIntVector,
-  DuckDBBigNumVector,
-  DuckDBBitVector,
-  DuckDBBlobVector,
-  DuckDBBooleanVector,
-  DuckDBConnection,
-  DuckDBDataChunk,
-  DuckDBDateValue,
-  DuckDBDateVector,
-  DuckDBDecimal128Vector,
-  DuckDBDecimal16Vector,
-  DuckDBDecimal32Vector,
-  DuckDBDecimal64Vector,
-  DuckDBDecimalValue,
-  DuckDBDoubleVector,
-  DuckDBEnum16Vector,
-  DuckDBEnum32Vector,
-  DuckDBEnum8Vector,
-  DuckDBFloatVector,
-  DuckDBGeometryType,
-  DuckDBHugeIntVector,
-  DuckDBInstance,
-  DuckDBIntegerVector,
-  DuckDBIntervalVector,
-  DuckDBListVector,
-  DuckDBMapVector,
-  DuckDBPendingResultState,
-  DuckDBResult,
-  DuckDBSmallIntVector,
-  DuckDBStructVector,
-  DuckDBTimeNSVector,
-  DuckDBTimeTZValue,
-  DuckDBTimeTZVector,
-  DuckDBTimeValue,
-  DuckDBTimeVector,
-  DuckDBTimestampMillisecondsVector,
-  DuckDBTimestampNanosecondsVector,
-  DuckDBTimestampSecondsVector,
-  DuckDBTimestampTZValue,
-  DuckDBTimestampTZVector,
-  DuckDBTimestampValue,
-  DuckDBTimestampVector,
-  DuckDBTinyIntVector,
-  DuckDBType,
-  DuckDBTypeId,
-  DuckDBUBigIntVector,
-  DuckDBUHugeIntVector,
-  DuckDBUIntegerVector,
-  DuckDBUSmallIntVector,
-  DuckDBUTinyIntVector,
-  DuckDBUUIDVector,
-  DuckDBUnionVector,
-  DuckDBValue,
-  DuckDBVarCharVector,
-  DuckDBVariantVector,
-  DuckDBVector,
-  ENUM,
-  FLOAT,
-  GEOMETRY,
-  HUGEINT,
-  INTEGER,
-  JSDuckDBValueConverter,
-  INTERVAL,
-  LIST,
-  MAP,
-  ResultReturnType,
-  SMALLINT,
-  SQLNULL,
-  STRUCT,
-  StatementType,
-  TIME,
-  TIMESTAMP,
-  TIMESTAMPTZ,
-  TIMESTAMP_MS,
-  TIMESTAMP_NS,
-  TIMESTAMP_S,
-  TIMETZ,
-  TINYINT,
-  TimeParts,
-  TimeTZParts,
-  TimestampParts,
-  UBIGINT,
-  UHUGEINT,
-  UINTEGER,
-  UNION,
-  USMALLINT,
-  UTINYINT,
-  UUID,
-  VARCHAR,
-  VARIANT,
-  DuckDBListValue,
-  DuckDBStructValue,
-  DuckDBVariantValue,
-  arrayValue,
-  bitValue,
-  blobValue,
-  configurationOptionDescriptions,
-  dateValue,
-  decimalValue,
-  geometryValue,
-  intervalValue,
-  listValue,
-  mapValue,
-  quotedIdentifier,
-  quotedString,
-  structValue,
-  timeTZValue,
-  timeValue,
-  timestampMillisValue,
-  timestampNanosValue,
-  timestampSecondsValue,
-  timestampValue,
-  unionValue,
-  uuidValue,
-  variantValue,
-  version
-} from '../src';
-import { varintDecode } from '../src/conversion/varintDecode';
-import { DuckDBInstanceCache } from '../src/DuckDBInstanceCache';
-import { DuckDBScalarFunction } from '../src/DuckDBScalarFunction';
-import { replaceSqlNullWithInteger } from './util/replaceSqlNullWithInteger';
-import {
-  ColumnNameAndType,
-  createTestAllTypesColumnNameAndTypeObjects,
-  createTestAllTypesColumnNameAndTypeObjectsJson,
-  createTestAllTypesColumnNamesAndTypesJson,
-  createTestAllTypesColumnTypes,
-  createTestAllTypesColumns,
-  createTestAllTypesColumnsJson,
-  createTestAllTypesColumnsObjectJson,
-  createTestAllTypesRowObjectsJson,
-  createTestAllTypesRowsJson,
-} from './util/testAllTypes';
-import {
-  createTestJSColumnsJS,
-  createTestJSColumnsObjectJS,
-  createTestJSQuery,
-  createTestJSRowObjectsJS,
-  createTestJSRowsJS,
-} from './util/testJS';
-
-async function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-async function withConnection(
-  fn: (connection: DuckDBConnection) => Promise<void>,
-) {
-  const instance = await DuckDBInstance.create();
-  const connection = await instance.connect();
-  await fn(connection);
-}
-
-function assertColumns(
-  result: DuckDBResult,
-  expectedColumns: readonly ColumnNameAndType[],
-) {
-  assert.strictEqual(
-    result.columnCount,
-    expectedColumns.length,
-    'column count',
-  );
-  for (let i = 0; i < expectedColumns.length; i++) {
-    const { name, type } = expectedColumns[i];
-    assert.strictEqual(result.columnName(i), name, 'column name');
-    assert.strictEqual(
-      result.columnTypeId(i),
-      type.typeId,
-      `column type id (column: ${name})`,
-    );
-    assert.deepStrictEqual(
-      result.columnType(i),
-      type,
-      `column type (column: ${name})`,
-    );
-  }
-}
-
-function isVectorType<
-  TValue extends DuckDBValue,
-  TVector extends DuckDBVector<TValue>,
->(
-  vector: DuckDBVector<any> | null,
-  vectorType: new (...args: any[]) => TVector,
-): vector is TVector {
-  return vector instanceof vectorType;
-}
-
-function getColumnVector<
-  TValue extends DuckDBValue,
-  TVector extends DuckDBVector<TValue>,
->(
-  chunk: DuckDBDataChunk,
-  columnIndex: number,
-  vectorType: new (...args: any[]) => TVector,
-): TVector {
-  const columnVector = chunk.getColumnVector(columnIndex);
-  if (!isVectorType<TValue, TVector>(columnVector, vectorType)) {
-    assert.fail(`expected column ${columnIndex} to be a ${vectorType}`);
-  }
-  return columnVector;
-}
-
-function assertVectorValues<TValue extends DuckDBValue>(
-  vector: DuckDBVector<TValue> | null | undefined,
-  values: readonly TValue[],
-  vectorName: string,
-) {
-  if (!vector) {
-    assert.fail(`${vectorName} unexpectedly null or undefined`);
-  }
-  assert.strictEqual(
-    vector.itemCount,
-    values.length,
-    `expected vector ${vectorName} item count to be ${values.length} but found ${vector.itemCount}`,
-  );
-  for (let i = 0; i < values.length; i++) {
-    const actual: TValue | null = vector.getItem(i);
-    const expected = values[i];
-    assert.deepStrictEqual(
-      actual,
-      expected,
-      `expected vector ${vectorName}[${i}] to be ${expected} but found ${actual}`,
-    );
-  }
-}
-
-function assertValues<
-  TValue extends DuckDBValue,
-  TVector extends DuckDBVector<TValue>,
->(
-  chunk: DuckDBDataChunk,
-  columnIndex: number,
-  vectorType: new (...args: any[]) => TVector,
-  values: readonly (TValue | null)[],
-) {
-  const vector = getColumnVector(chunk, columnIndex, vectorType);
-  assertVectorValues(vector, values, `${columnIndex}`);
-}
-
-function bigints(start: bigint, end: bigint) {
-  return Array.from({ length: Number(end - start) + 1 }).map(
-    (_, i) => start + BigInt(i),
-  );
-}
-
-describe('api', () => {
-  beforeAll(() => {
-    // Set this to a constant so tests don't depend on the local timezone.
-    DuckDBTimestampTZValue.timezoneOffsetInMinutes = -330;
-  });
-  test('should expose version', () => {
-    const ver = version();
-    assert.ok(ver.startsWith('v'), `version starts with 'v'`);
-  });
-  test('should expose configuration option descriptions', () => {
-    const descriptions = configurationOptionDescriptions();
-    assert.ok(descriptions['memory_limit'], `descriptions has 'memory_limit'`);
-  });
-  test('ReturnResultType enum', () => {
-    assert.equal(ResultReturnType.INVALID, 0);
-    assert.equal(ResultReturnType.CHANGED_ROWS, 1);
-    assert.equal(ResultReturnType.NOTHING, 2);
-    assert.equal(ResultReturnType.QUERY_RESULT, 3);
-
-    assert.equal(ResultReturnType[ResultReturnType.INVALID], 'INVALID');
-    assert.equal(
-      ResultReturnType[ResultReturnType.CHANGED_ROWS],
-      'CHANGED_ROWS',
-    );
-    assert.equal(ResultReturnType[ResultReturnType.NOTHING], 'NOTHING');
-    assert.equal(
-      ResultReturnType[ResultReturnType.QUERY_RESULT],
-      'QUERY_RESULT',
-    );
-  });
-  test('StatementType enum', () => {
-    assert.equal(StatementType.INVALID, 0);
-    assert.equal(StatementType.SELECT, 1);
-    assert.equal(StatementType.INSERT, 2);
-    assert.equal(StatementType.UPDATE, 3);
-    assert.equal(StatementType.EXPLAIN, 4);
-    assert.equal(StatementType.DELETE, 5);
-    assert.equal(StatementType.PREPARE, 6);
-    assert.equal(StatementType.CREATE, 7);
-    assert.equal(StatementType.EXECUTE, 8);
-    assert.equal(StatementType.ALTER, 9);
-    assert.equal(StatementType.TRANSACTION, 10);
-    assert.equal(StatementType.COPY, 11);
-    assert.equal(StatementType.ANALYZE, 12);
-    assert.equal(StatementType.VARIABLE_SET, 13);
-    assert.equal(StatementType.CREATE_FUNC, 14);
-    assert.equal(StatementType.DROP, 15);
-    assert.equal(StatementType.EXPORT, 16);
-    assert.equal(StatementType.PRAGMA, 17);
-    assert.equal(StatementType.VACUUM, 18);
-    assert.equal(StatementType.CALL, 19);
-    assert.equal(StatementType.SET, 20);
-    assert.equal(StatementType.LOAD, 21);
-    assert.equal(StatementType.RELATION, 22);
-    assert.equal(StatementType.EXTENSION, 23);
-    assert.equal(StatementType.LOGICAL_PLAN, 24);
-    assert.equal(StatementType.ATTACH, 25);
-    assert.equal(StatementType.DETACH, 26);
-    assert.equal(StatementType.MULTI, 27);
-
-    assert.equal(StatementType[StatementType.INVALID], 'INVALID');
-    assert.equal(StatementType[StatementType.SELECT], 'SELECT');
-    assert.equal(StatementType[StatementType.INSERT], 'INSERT');
-    assert.equal(StatementType[StatementType.UPDATE], 'UPDATE');
-    assert.equal(StatementType[StatementType.EXPLAIN], 'EXPLAIN');
-    assert.equal(StatementType[StatementType.DELETE], 'DELETE');
-    assert.equal(StatementType[StatementType.PREPARE], 'PREPARE');
-    assert.equal(StatementType[StatementType.CREATE], 'CREATE');
-    assert.equal(StatementType[StatementType.EXECUTE], 'EXECUTE');
-    assert.equal(StatementType[StatementType.ALTER], 'ALTER');
-    assert.equal(StatementType[StatementType.TRANSACTION], 'TRANSACTION');
-    assert.equal(StatementType[StatementType.COPY], 'COPY');
-    assert.equal(StatementType[StatementType.ANALYZE], 'ANALYZE');
-    assert.equal(StatementType[StatementType.VARIABLE_SET], 'VARIABLE_SET');
-    assert.equal(StatementType[StatementType.CREATE_FUNC], 'CREATE_FUNC');
-    assert.equal(StatementType[StatementType.DROP], 'DROP');
-    assert.equal(StatementType[StatementType.EXPORT], 'EXPORT');
-    assert.equal(StatementType[StatementType.PRAGMA], 'PRAGMA');
-    assert.equal(StatementType[StatementType.VACUUM], 'VACUUM');
-    assert.equal(StatementType[StatementType.CALL], 'CALL');
-    assert.equal(StatementType[StatementType.SET], 'SET');
-    assert.equal(StatementType[StatementType.LOAD], 'LOAD');
-    assert.equal(StatementType[StatementType.RELATION], 'RELATION');
-    assert.equal(StatementType[StatementType.EXTENSION], 'EXTENSION');
-    assert.equal(StatementType[StatementType.LOGICAL_PLAN], 'LOGICAL_PLAN');
-    assert.equal(StatementType[StatementType.ATTACH], 'ATTACH');
-    assert.equal(StatementType[StatementType.DETACH], 'DETACH');
-    assert.equal(StatementType[StatementType.MULTI], 'MULTI');
-  });
-  test('DuckDBType toString', () => {
-    assert.equal(BOOLEAN.toString(), 'BOOLEAN');
-    assert.equal(TINYINT.toString(), 'TINYINT');
-    assert.equal(SMALLINT.toString(), 'SMALLINT');
-    assert.equal(INTEGER.toString(), 'INTEGER');
-    assert.equal(BIGINT.toString(), 'BIGINT');
-    assert.equal(UTINYINT.toString(), 'UTINYINT');
-    assert.equal(USMALLINT.toString(), 'USMALLINT');
-    assert.equal(UINTEGER.toString(), 'UINTEGER');
-    assert.equal(UBIGINT.toString(), 'UBIGINT');
-    assert.equal(FLOAT.toString(), 'FLOAT');
-    assert.equal(DOUBLE.toString(), 'DOUBLE');
-    assert.equal(TIMESTAMP.toString(), 'TIMESTAMP');
-    assert.equal(DATE.toString(), 'DATE');
-    assert.equal(TIME.toString(), 'TIME');
-    assert.equal(INTERVAL.toString(), 'INTERVAL');
-    assert.equal(HUGEINT.toString(), 'HUGEINT');
-    assert.equal(UHUGEINT.toString(), 'UHUGEINT');
-    assert.equal(VARCHAR.toString(), 'VARCHAR');
-    assert.equal(BLOB.toString(), 'BLOB');
-    assert.equal(DECIMAL(17, 5).toString(), 'DECIMAL(17,5)');
-    assert.equal(TIMESTAMP_S.toString(), 'TIMESTAMP_S');
-    assert.equal(TIMESTAMP_MS.toString(), 'TIMESTAMP_MS');
-    assert.equal(TIMESTAMP_NS.toString(), 'TIMESTAMP_NS');
-    assert.equal(
-      ENUM(['fly', 'swim', 'walk']).toString(),
-      `ENUM('fly', 'swim', 'walk')`,
-    );
-    assert.equal(LIST(INTEGER).toString(), 'INTEGER[]');
-    assert.equal(
-      STRUCT({ 'id': VARCHAR, 'ts': TIMESTAMP }).toString(),
-      'STRUCT("id" VARCHAR, "ts" TIMESTAMP)',
-    );
-    assert.equal(MAP(INTEGER, VARCHAR).toString(), 'MAP(INTEGER, VARCHAR)');
-    assert.equal(ARRAY(INTEGER, 3).toString(), 'INTEGER[3]');
-    assert.equal(UUID.toString(), 'UUID');
-    assert.equal(
-      UNION({ 'str': VARCHAR, 'num': INTEGER }).toString(),
-      'UNION("str" VARCHAR, "num" INTEGER)',
-    );
-    assert.equal(BIT.toString(), 'BIT');
-    assert.equal(TIMETZ.toString(), 'TIME WITH TIME ZONE');
-    assert.equal(TIMESTAMPTZ.toString(), 'TIMESTAMP WITH TIME ZONE');
-    assert.equal(ANY.toString(), 'ANY');
-    assert.equal(BIGNUM.toString(), 'BIGNUM');
-    assert.equal(SQLNULL.toString(), 'SQLNULL');
-  });
-  test('should support creating, connecting, running a basic query, and reading results', async () => {
-    const instance = await DuckDBInstance.create();
-    const connection = await instance.connect();
-    const result = await connection.run('select 42 as num');
-    assertColumns(result, [{ name: 'num', type: INTEGER }]);
-    const chunk = await result.fetchChunk();
-    assert.isDefined(chunk);
-    if (chunk) {
-      assert.strictEqual(chunk.columnCount, 1);
-      assert.strictEqual(chunk.rowCount, 1);
-      assertValues<number, DuckDBIntegerVector>(
-        chunk,
-        0,
-        DuckDBIntegerVector,
-        [42],
-      );
-    }
-    instance.closeSync();
-  });
-  test('disconnecting connections', async () => {
-    const instance = await DuckDBInstance.create();
-    const connection = await instance.connect();
-    const prepared1 = await connection.prepare('select 1');
-    assert.isDefined(prepared1);
-    prepared1.destroySync();
-    connection.disconnectSync();
-    try {
-      await connection.prepare('select 2');
-      assert.fail('should throw');
-    } catch (err) {
-      assert.deepEqual(
-        err,
-        new Error('Failed to prepare: connection disconnected'),
-      );
-    }
-    // ensure double-disconnect doesn't break anything
-    connection.disconnectSync();
-  });
-  test('client context', async () => {
-    await withConnection(async (connection) => {
-      const clientContext = connection.clientContext;
-      assert.isDefined(clientContext);
-      assert.isDefined(clientContext.connectionId);
-    });
-  });
-  test('should support running prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const params: ColumnNameAndType[] = [
-        { name: 'boolean', type: BOOLEAN },
-        { name: 'integer', type: INTEGER },
-        { name: 'varchar', type: VARCHAR },
-        { name: 'timestamp_s', type: TIMESTAMP_S },
-        { name: 'timestamp_ms', type: TIMESTAMP_MS },
-        { name: 'timestamp_ns', type: TIMESTAMP_NS },
-        { name: 'enum', type: ENUM(['fly', 'swim', 'walk']) },
-        { name: 'list_int', type: LIST(INTEGER) },
-        { name: 'list_dec', type: LIST(DECIMAL(4, 1)) },
-        { name: 'list_null', type: LIST(SQLNULL) },
-        { name: 'struct', type: STRUCT({ 'a': INTEGER, 'b': VARCHAR }) },
-        { name: 'array', type: ARRAY(INTEGER, 3) },
-        { name: 'map', type: MAP(INTEGER, VARCHAR) },
-        { name: 'union', type: UNION({ 'name': VARCHAR, 'age': SMALLINT }) },
-        { name: 'uuid', type: UUID },
-        { name: 'bit', type: BIT },
-        { name: 'timetz', type: TIMETZ },
-        { name: 'timestamptz', type: TIMESTAMPTZ },
-        { name: 'bignum', type: BIGNUM },
-        { name: 'null_value', type: SQLNULL },
-      ];
-
-      const sql = `select ${params
-        .map((p) => `$${p.name} as ${p.name}`)
-        .join(', ')}`;
-      const prepared = await connection.prepare(sql);
-
-      assert.strictEqual(prepared.parameterCount, params.length);
-      for (let i = 0; i < params.length; i++) {
-        assert.strictEqual(
-          prepared.parameterName(i + 1),
-          params[i].name,
-          `param ${i} name mismatch`,
-        );
-      }
-
-      let i = 1;
-      prepared.bindBoolean(i++, true);
-      prepared.bindInteger(i++, 10);
-      prepared.bindVarchar(i++, 'abc');
-      prepared.bindTimestampSeconds(i++, TIMESTAMP_S.max);
-      prepared.bindTimestampMilliseconds(i++, TIMESTAMP_MS.max);
-      prepared.bindTimestampNanoseconds(i++, TIMESTAMP_NS.max);
-      prepared.bindEnum(i++, 'swim', ENUM(['fly', 'swim', 'walk']));
-      prepared.bindList(i++, [100, 200, 300]);
-      prepared.bindList(
-        i++,
-        [decimalValue(9876n, 4, 1), decimalValue(5432n, 4, 1)],
-        LIST(DECIMAL(4, 1)),
-      );
-      prepared.bindList(i++, [null]);
-      prepared.bindStruct(i++, { 'a': 42, 'b': 'duck' });
-      prepared.bindArray(i++, [100, 200, 300]);
-      prepared.bindMap(
-        i++,
-        mapValue([
-          { key: 100, value: 'swim' },
-          { key: 101, value: 'walk' },
-          { key: 102, value: 'fly' },
-        ]),
-      );
-      (prepared.bindUnion(
-        i++,
-        unionValue('age', 42),
-        UNION({ 'name': VARCHAR, 'age': SMALLINT }),
-      ),
-        prepared.bindUUID(i++, uuidValue(0xf0e1d2c3b4a596870123456789abcdefn)));
-      prepared.bindBit(i++, bitValue('0010001001011100010101011010111'));
-      prepared.bindTimeTZ(i++, TIMETZ.max);
-      prepared.bindTimestampTZ(i++, TIMESTAMPTZ.max);
-      prepared.bindBigNum(i++, BIGNUM.max);
-      prepared.bindNull(i++);
-
-      for (let i = 0; i < params.length; i++) {
-        let type = params[i].type;
-        if (type.typeId === DuckDBTypeId.VARCHAR) {
-          // VARCHAR type is reported incorrectly; see https://github.com/duckdb/duckdb/issues/16137
-          continue;
-        }
-        assert.equal(
-          prepared.parameterTypeId(i + 1),
-          type.typeId,
-          `param ${i} type id mismatch`,
-        );
-        assert.deepEqual(
-          prepared.parameterType(i + 1),
-          type,
-          `param ${i} type mismatch`,
-        );
-      }
-
-      const result = await prepared.run();
-
-      // In the result, SQLNULL params get type INTEGER.
-      const expectedColumns = params.map((p) => {
-        const replacedType = replaceSqlNullWithInteger(p.type);
-        if (replacedType !== p.type) {
-          return { ...p, type: replacedType };
-        }
-        return p;
-      });
-
-      assertColumns(result, expectedColumns);
-
-      const chunk = await result.fetchChunk();
-
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, expectedColumns.length);
-        assert.strictEqual(chunk.rowCount, 1);
-        let i = 0;
-        assertValues<boolean, DuckDBBooleanVector>(
-          chunk,
-          i++,
-          DuckDBBooleanVector,
-          [true],
-        );
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          i++,
-          DuckDBIntegerVector,
-          [10],
-        );
-        assertValues<string, DuckDBVarCharVector>(
-          chunk,
-          i++,
-          DuckDBVarCharVector,
-          ['abc'],
-        );
-        assertValues(chunk, i++, DuckDBTimestampSecondsVector, [
-          TIMESTAMP_S.max,
-        ]);
-        assertValues(chunk, i++, DuckDBTimestampMillisecondsVector, [
-          TIMESTAMP_MS.max,
-        ]);
-        assertValues(chunk, i++, DuckDBTimestampNanosecondsVector, [
-          TIMESTAMP_NS.max,
-        ]);
-        assertValues<string, DuckDBEnum8Vector>(chunk, i++, DuckDBEnum8Vector, [
-          'swim',
-        ]);
-        assertValues(chunk, i++, DuckDBListVector, [
-          listValue([100, 200, 300]),
-        ]);
-        assertValues(chunk, i++, DuckDBListVector, [
-          listValue([decimalValue(9876n, 4, 1), decimalValue(5432n, 4, 1)]),
-        ]);
-        assertValues(chunk, i++, DuckDBListVector, [listValue([null])]);
-        assertValues(chunk, i++, DuckDBStructVector, [
-          structValue({ 'a': 42, 'b': 'duck' }),
-        ]);
-        assertValues(chunk, i++, DuckDBArrayVector, [
-          arrayValue([100, 200, 300]),
-        ]);
-        assertValues(chunk, i++, DuckDBMapVector, [
-          mapValue([
-            { key: 100, value: 'swim' },
-            { key: 101, value: 'walk' },
-            { key: 102, value: 'fly' },
-          ]),
-        ]);
-        assertValues(chunk, i++, DuckDBUnionVector, [unionValue('age', 42)]);
-        assertValues(chunk, i++, DuckDBUUIDVector, [
-          uuidValue(0xf0e1d2c3b4a596870123456789abcdefn),
-        ]);
-        assertValues(chunk, i++, DuckDBBitVector, [
-          bitValue('0010001001011100010101011010111'),
-        ]);
-        assertValues(chunk, i++, DuckDBTimeTZVector, [TIMETZ.max]);
-        assertValues(chunk, i++, DuckDBTimestampTZVector, [TIMESTAMPTZ.max]);
-        assertValues(chunk, i++, DuckDBBigNumVector, [BIGNUM.max]);
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          i++,
-          DuckDBIntegerVector,
-          [null],
-        );
-      }
-    });
-  });
-  test('should support prepare statement bind with list', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare(
-        'select $1 as a, $2 as b, $3 as c',
-      );
-      prepared.bind([42, 'duck', listValue([10, 11, 12])]);
-      const result = await prepared.run();
-      assertColumns(result, [
-        { name: 'a', type: INTEGER },
-        { name: 'b', type: VARCHAR },
-        { name: 'c', type: LIST(INTEGER) },
-      ]);
-      const chunk = await result.fetchChunk();
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, 3);
-        assert.strictEqual(chunk.rowCount, 1);
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          0,
-          DuckDBIntegerVector,
-          [42],
-        );
-        assertValues<string, DuckDBVarCharVector>(
-          chunk,
-          1,
-          DuckDBVarCharVector,
-          ['duck'],
-        );
-        assertValues(chunk, 2, DuckDBListVector, [listValue([10, 11, 12])]);
-      }
-    });
-  });
-  test('should support prepare statement bind with object', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare(
-        'select $a as a, $b as b, $c as c, $d as d, $e as e, $f as f',
-      );
-      prepared.bind(
-        {
-          a: 42,
-          b: 42.3,
-          c: 'duck',
-          d: listValue([10, 11, 12]),
-          e: arrayValue([10.1, 11.2, 12.3]),
-          f: arrayValue([10, 11, 12]),
-        },
-        {
-          f: ARRAY(FLOAT, 2),
-        },
-      );
-      const result = await prepared.run();
-      assertColumns(result, [
-        { name: 'a', type: INTEGER },
-        { name: 'b', type: DOUBLE },
-        { name: 'c', type: VARCHAR },
-        { name: 'd', type: LIST(INTEGER) },
-        { name: 'e', type: ARRAY(DOUBLE, 3) },
-        { name: 'f', type: ARRAY(FLOAT, 3) },
-      ]);
-      const chunk = await result.fetchChunk();
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, 6);
-        assert.strictEqual(chunk.rowCount, 1);
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          0,
-          DuckDBIntegerVector,
-          [42],
-        );
-        assertValues<number, DuckDBDoubleVector>(
-          chunk,
-          1,
-          DuckDBDoubleVector,
-          [42.3],
-        );
-        assertValues<string, DuckDBVarCharVector>(
-          chunk,
-          2,
-          DuckDBVarCharVector,
-          ['duck'],
-        );
-        assertValues(chunk, 3, DuckDBListVector, [listValue([10, 11, 12])]);
-        assertValues(chunk, 4, DuckDBArrayVector, [
-          arrayValue([10.1, 11.2, 12.3]),
-        ]);
-        assertValues(chunk, 5, DuckDBArrayVector, [arrayValue([10, 11, 12])]);
-      }
-    });
-  });
-  test('should fail gracefully when binding structs contain ANY types to prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare('select ?');
-      try {
-        prepared.bindStruct(
-          0,
-          structValue({ 'a': null }),
-          STRUCT({ 'a': ANY }),
-        );
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(
-          err,
-          new Error(
-            'Cannot create structs with an entry type of ANY. Specify a specific type.',
-          ),
-        );
-      }
-    });
-  });
-  test('should fail gracefully when type cannot be inferred when binding lists to prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare('select ?');
-      try {
-        prepared.bind([listValue([])]);
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(
-          err,
-          new Error(
-            'Cannot create lists with item type of ANY. Specify a specific type.',
-          ),
-        );
-      }
-    });
-  });
-  test('should fail gracefully when type cannot be inferred when binding arrays to prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare('select ?');
-      try {
-        prepared.bind([arrayValue([])]);
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(
-          err,
-          new Error(
-            'Cannot create arrays with item type of ANY. Specify a specific type.',
-          ),
-        );
-      }
-    });
-  });
-  test('should infer integer and floating-point values when binding to prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare('select ? as i1, ? as i2, ? as b1, ? as b2, ? as d');
-      prepared.bind([2_147_483_647, -2_147_483_648, 2_147_483_648, -2_147_483_649, 3.14]);
-      const result = await prepared.run();
-      assertColumns(result, [
-        { name: 'i1', type: INTEGER },
-        { name: 'i2', type: INTEGER },
-        { name: 'b1', type: BIGINT },
-        { name: 'b2', type: BIGINT },
-        { name: 'd', type: DOUBLE },
-      ]);
-      const chunk = await result.fetchChunk();
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, 5);
-        assert.strictEqual(chunk.rowCount, 1);
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          0,
-          DuckDBIntegerVector,
-          [2_147_483_647],
-        );
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          1,
-          DuckDBIntegerVector,
-          [-2_147_483_648],
-        );
-        assertValues<bigint, DuckDBBigIntVector>(
-          chunk,
-          2,
-          DuckDBBigIntVector,
-          [2_147_483_648n],
-        );
-        assertValues<bigint, DuckDBBigIntVector>(
-          chunk,
-          3,
-          DuckDBBigIntVector,
-          [-2_147_483_649n],
-        );
-        assertValues<number, DuckDBDoubleVector>(
-          chunk,
-          4,
-          DuckDBDoubleVector,
-          [3.14],
-        );
-      }
-    });
-  });
-  test('should support starting prepared statements and running them incrementally', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare(
-        'select int from test_all_types()',
-      );
-      const pending = prepared.start();
-      let taskCount = 0;
-      while (pending.runTask() !== DuckDBPendingResultState.RESULT_READY) {
-        taskCount++;
-        // arbitrary upper bound on the number of tasks expected for this simple query
-        if (taskCount > 100) {
-          assert.fail('Unexpectedly large number of tasks');
-        }
-        await sleep(1);
-      }
-      // console.debug('task count: ', taskCount);
-      const result = await pending.getResult();
-      assertColumns(result, [{ name: 'int', type: INTEGER }]);
-      const chunk = await result.fetchChunk();
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, 1);
-        assert.strictEqual(chunk.rowCount, 3);
-        assertValues(chunk, 0, DuckDBIntegerVector, [
-          INTEGER.min,
-          INTEGER.max,
-          null,
-        ]);
-      }
-    });
-  });
-  test('should support streaming results from prepared statements', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare('from range(10000)');
-      const pending = prepared.start();
-      const result = await pending.getResult();
-      assertColumns(result, [{ name: 'range', type: BIGINT }]);
-      const chunks: DuckDBDataChunk[] = [];
-      let currentChunk: DuckDBDataChunk | null = null;
-      currentChunk = await result.fetchChunk();
-      while (currentChunk && currentChunk.rowCount > 0) {
-        chunks.push(currentChunk);
-        currentChunk = await result.fetchChunk();
-      }
-      currentChunk = null;
-      assert.strictEqual(chunks.length, 5); // ceil(10000 / 2048) = 5
-      assertValues(chunks[0], 0, DuckDBBigIntVector, bigints(0n, 2048n - 1n));
-      assertValues(
-        chunks[1],
-        0,
-        DuckDBBigIntVector,
-        bigints(2048n, 2048n * 2n - 1n),
-      );
-      assertValues(
-        chunks[2],
-        0,
-        DuckDBBigIntVector,
-        bigints(2048n * 2n, 2048n * 3n - 1n),
-      );
-      assertValues(
-        chunks[3],
-        0,
-        DuckDBBigIntVector,
-        bigints(2048n * 3n, 2048n * 4n - 1n),
-      );
-      assertValues(
-        chunks[4],
-        0,
-        DuckDBBigIntVector,
-        bigints(2048n * 4n, 9999n),
-      );
-    });
-  });
-  test('prepared statement column info (valid)', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare(
-        'select $1::INTEGER as a, $2::VARCHAR as b, $3::INTEGER[] as c',
-      );
-      assert.equal(prepared.columnCount, 3);
-      assert.equal(prepared.columnName(0), 'a');
-      assert.equal(prepared.columnTypeId(0), DuckDBTypeId.INTEGER);
-      assert.deepEqual(prepared.columnType(0), INTEGER);
-      assert.equal(prepared.columnName(1), 'b');
-      assert.equal(prepared.columnTypeId(1), DuckDBTypeId.VARCHAR);
-      assert.deepEqual(prepared.columnType(1), VARCHAR);
-      assert.equal(prepared.columnName(2), 'c');
-      assert.equal(prepared.columnTypeId(2), DuckDBTypeId.LIST);
-      assert.deepEqual(prepared.columnType(2), LIST(INTEGER));
-    });
-  });
-  test('prepared statement column info (invalid)', async () => {
-    await withConnection(async (connection) => {
-      const prepared = await connection.prepare(
-        'select $1::INTEGER as a, $2::VARCHAR as b, $3 as c',
-      );
-      // When any column types are ambiguous, DuckDB returns a column count of 1 with a type of INVALID.
-      assert.equal(prepared.columnCount, 1);
-      assert.equal(prepared.columnTypeId(0), DuckDBTypeId.INVALID);
-    });
-  });
-  test('runAndReadAll with params', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll('select ? as n', [17]);
-      const rows = reader.getRowObjects();
-      assert.deepEqual(rows, [{ n: 17 }]);
-    });
-  });
-  test('should support PIVOT (no parameters)', async () => {
-    await withConnection(async (connection) => {
-      await connection.run(`
-CREATE TABLE cities (
-    country VARCHAR, name VARCHAR, year INTEGER, population INTEGER
-);
-INSERT INTO cities VALUES
-    ('NL', 'Amsterdam', 2000, 1005),
-    ('NL', 'Amsterdam', 2010, 1065),
-    ('NL', 'Amsterdam', 2020, 1158),
-    ('US', 'Seattle', 2000, 564),
-    ('US', 'Seattle', 2010, 608),
-    ('US', 'Seattle', 2020, 738),
-    ('US', 'New York City', 2000, 8015),
-    ('US', 'New York City', 2010, 8175),
-    ('US', 'New York City', 2020, 8772);
-    `);
-      const reader = await connection.runAndReadAll(`
-FROM (
-  PIVOT cities
-  ON year
-  USING sum(population)
-)
-ORDER BY name
-      `);
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        country: ['NL', 'US', 'US'],
-        name: ['Amsterdam', 'New York City', 'Seattle'],
-        '2000': [1005n, 8015n, 564n],
-        '2010': [1065n, 8175n, 608n],
-        '2020': [1158n, 8772n, 738n],
-      });
-    });
-  });
-  test('should support PIVOT (with parameters)', async () => {
-    await withConnection(async (connection) => {
-      await connection.run(`
-CREATE TABLE cities (
-    country VARCHAR, name VARCHAR, year INTEGER, population INTEGER
-);
-INSERT INTO cities VALUES
-    ('NL', 'Amsterdam', 2000, 1005),
-    ('NL', 'Amsterdam', 2010, 1065),
-    ('NL', 'Amsterdam', 2020, 1158),
-    ('US', 'Seattle', 2000, 564),
-    ('US', 'Seattle', 2010, 608),
-    ('US', 'Seattle', 2020, 738),
-    ('US', 'New York City', 2000, 8015),
-    ('US', 'New York City', 2010, 8175),
-    ('US', 'New York City', 2020, 8772);
-    `);
-      const reader = await connection.runAndReadAll(
-        `
-FROM (
-  PIVOT cities
-  ON year
-  USING sum(population)
-)
-WHERE country = $country
-ORDER BY name
-      `,
-        { country: 'US' },
-      );
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        country: ['US', 'US'],
-        name: ['New York City', 'Seattle'],
-        '2000': [8015n, 564n],
-        '2010': [8175n, 608n],
-        '2020': [8772n, 738n],
-      });
-    });
-  });
-  test('should support all data types', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.run(
-        'from test_all_types(use_large_enum=true)',
-      );
-      assertColumns(result, createTestAllTypesColumnNameAndTypeObjects());
-
-      const chunk = await result.fetchChunk();
-      assert.isDefined(chunk);
-      if (chunk) {
-        assert.strictEqual(chunk.columnCount, 56);
-        assert.strictEqual(chunk.rowCount, 3);
-
-        const testAllTypesColumns = createTestAllTypesColumns();
-
-        assertValues(chunk, 0, DuckDBBooleanVector, testAllTypesColumns[0]);
-        assertValues(chunk, 1, DuckDBTinyIntVector, testAllTypesColumns[1]);
-        assertValues(chunk, 2, DuckDBSmallIntVector, testAllTypesColumns[2]);
-        assertValues(chunk, 3, DuckDBIntegerVector, testAllTypesColumns[3]);
-        assertValues(chunk, 4, DuckDBBigIntVector, testAllTypesColumns[4]);
-        assertValues(chunk, 5, DuckDBHugeIntVector, testAllTypesColumns[5]);
-        assertValues(chunk, 6, DuckDBUHugeIntVector, testAllTypesColumns[6]);
-        assertValues(chunk, 7, DuckDBUTinyIntVector, testAllTypesColumns[7]);
-        assertValues(chunk, 8, DuckDBUSmallIntVector, testAllTypesColumns[8]);
-        assertValues(chunk, 9, DuckDBUIntegerVector, testAllTypesColumns[9]);
-        assertValues(chunk, 10, DuckDBUBigIntVector, testAllTypesColumns[10]);
-        assertValues(chunk, 11, DuckDBBigNumVector, testAllTypesColumns[11]);
-        assertValues(chunk, 12, DuckDBDateVector, testAllTypesColumns[12]);
-        assertValues(chunk, 13, DuckDBTimeVector, testAllTypesColumns[13]);
-        assertValues(chunk, 14, DuckDBTimestampVector, testAllTypesColumns[14]);
-        assertValues(
-          chunk,
-          15,
-          DuckDBTimestampSecondsVector,
-          testAllTypesColumns[15],
-        );
-        assertValues(
-          chunk,
-          16,
-          DuckDBTimestampMillisecondsVector,
-          testAllTypesColumns[16],
-        );
-        assertValues(
-          chunk,
-          17,
-          DuckDBTimestampNanosecondsVector,
-          testAllTypesColumns[17],
-        );
-        assertValues(chunk, 18, DuckDBTimeTZVector, testAllTypesColumns[18]);
-        assertValues(
-          chunk,
-          19,
-          DuckDBTimestampTZVector,
-          testAllTypesColumns[19],
-        );
-        assertValues(chunk, 20, DuckDBFloatVector, testAllTypesColumns[20]);
-        assertValues(chunk, 21, DuckDBDoubleVector, testAllTypesColumns[21]);
-        assertValues(chunk, 22, DuckDBDecimal16Vector, testAllTypesColumns[22]);
-        assertValues(chunk, 23, DuckDBDecimal32Vector, testAllTypesColumns[23]);
-        assertValues(chunk, 24, DuckDBDecimal64Vector, testAllTypesColumns[24]);
-        assertValues(
-          chunk,
-          25,
-          DuckDBDecimal128Vector,
-          testAllTypesColumns[25],
-        );
-        assertValues(chunk, 26, DuckDBUUIDVector, testAllTypesColumns[26]);
-        assertValues(chunk, 27, DuckDBIntervalVector, testAllTypesColumns[27]);
-        assertValues(chunk, 28, DuckDBVarCharVector, testAllTypesColumns[28]);
-        assertValues(chunk, 29, DuckDBBlobVector, testAllTypesColumns[29]);
-        assertValues(chunk, 30, DuckDBBitVector, testAllTypesColumns[30]);
-        assertValues(chunk, 31, DuckDBEnum8Vector, testAllTypesColumns[31]);
-        assertValues(chunk, 32, DuckDBEnum16Vector, testAllTypesColumns[32]);
-        assertValues(chunk, 33, DuckDBEnum32Vector, testAllTypesColumns[33]);
-        // int_array
-        assertValues(chunk, 34, DuckDBListVector, testAllTypesColumns[34]);
-        // double_array
-        assertValues(chunk, 35, DuckDBListVector, testAllTypesColumns[35]);
-        // date_array
-        assertValues(chunk, 36, DuckDBListVector, testAllTypesColumns[36]);
-        // timestamp_array
-        assertValues(chunk, 37, DuckDBListVector, testAllTypesColumns[37]);
-        // timestamptz_array
-        assertValues(chunk, 38, DuckDBListVector, testAllTypesColumns[38]);
-        // varchar_array
-        assertValues(chunk, 39, DuckDBListVector, testAllTypesColumns[39]);
-        // nested_int_array
-        assertValues(chunk, 40, DuckDBListVector, testAllTypesColumns[40]);
-        assertValues(chunk, 41, DuckDBStructVector, testAllTypesColumns[41]);
-        // struct_of_arrays
-        assertValues(chunk, 42, DuckDBStructVector, testAllTypesColumns[42]);
-        // array_of_structs
-        assertValues(chunk, 43, DuckDBListVector, testAllTypesColumns[43]);
-        assertValues(chunk, 44, DuckDBMapVector, testAllTypesColumns[44]);
-        assertValues(chunk, 45, DuckDBUnionVector, testAllTypesColumns[45]);
-        // fixed_int_array
-        assertValues(chunk, 46, DuckDBArrayVector, testAllTypesColumns[46]);
-        // fixed_varchar_array
-        assertValues(chunk, 47, DuckDBArrayVector, testAllTypesColumns[47]);
-        // fixed_nested_int_array
-        assertValues(chunk, 48, DuckDBArrayVector, testAllTypesColumns[48]);
-        // fixed_nested_varchar_array
-        assertValues(chunk, 49, DuckDBArrayVector, testAllTypesColumns[49]);
-        // fixed_struct_array
-        assertValues(chunk, 50, DuckDBArrayVector, testAllTypesColumns[50]);
-        // struct_of_fixed_array
-        assertValues(chunk, 51, DuckDBStructVector, testAllTypesColumns[51]);
-        // fixed_array_of_int_list
-        assertValues(chunk, 52, DuckDBArrayVector, testAllTypesColumns[52]);
-        // list_of_fixed_int_array
-        assertValues(chunk, 53, DuckDBListVector, testAllTypesColumns[53]);
-        // time_ns
-        assertValues(chunk, 54, DuckDBTimeNSVector, testAllTypesColumns[54]);
-      }
-    });
-  });
-  test('values toString', () => {
-    // array
-    assert.equal(arrayValue([]).toString(), '[]');
-    assert.equal(arrayValue([1, 2, 3]).toString(), '[1, 2, 3]');
-    assert.equal(arrayValue(['a', 'b', 'c']).toString(), `['a', 'b', 'c']`);
-
-    // bit
-    assert.equal(bitValue('').toString(), '');
-    assert.equal(bitValue('10101').toString(), '10101');
-    assert.equal(
-      bitValue('0010001001011100010101011010111').toString(),
-      '0010001001011100010101011010111',
-    );
-
-    // blob
-    assert.equal(blobValue('').toString(), '');
-    assert.equal(
-      blobValue('thisisalongblob\x00withnullbytes').toString(),
-      'thisisalongblob\\x00withnullbytes',
-    );
-    assert.equal(blobValue('\x00\x00\x00a').toString(), '\\x00\\x00\\x00a');
-
-    // date
-    assert.equal(DATE.epoch.toString(), '1970-01-01');
-    assert.equal(DATE.max.toString(), '5881580-07-10');
-    assert.equal(DATE.min.toString(), '5877642-06-25 (BC)');
-
-    // decimal
-    assert.equal(decimalValue(0n, 4, 1).toString(), '0.0');
-    assert.equal(decimalValue(9876n, 4, 1).toString(), '987.6');
-    assert.equal(decimalValue(-9876n, 4, 1).toString(), '-987.6');
-
-    assert.equal(decimalValue(0n, 9, 4).toString(), '0.0000');
-    assert.equal(decimalValue(987654321n, 9, 4).toString(), '98765.4321');
-    assert.equal(decimalValue(-987654321n, 9, 4).toString(), '-98765.4321');
-
-    assert.equal(decimalValue(0n, 18, 6).toString(), '0.000000');
-    assert.equal(
-      decimalValue(987654321098765432n, 18, 6).toString(),
-      '987654321098.765432',
-    );
-    assert.equal(
-      decimalValue(-987654321098765432n, 18, 6).toString(),
-      '-987654321098.765432',
-    );
-
-    assert.equal(decimalValue(0n, 38, 10).toString(), '0.0000000000');
-    assert.equal(
-      decimalValue(98765432109876543210987654321098765432n, 38, 10).toString(),
-      '9876543210987654321098765432.1098765432',
-    );
-    assert.equal(
-      decimalValue(-98765432109876543210987654321098765432n, 38, 10).toString(),
-      '-9876543210987654321098765432.1098765432',
-    );
-
-    // interval
-    assert.equal(intervalValue(0, 0, 0n).toString(), '00:00:00');
-
-    assert.equal(intervalValue(1, 0, 0n).toString(), '1 month');
-    assert.equal(intervalValue(-1, 0, 0n).toString(), '-1 month');
-    assert.equal(intervalValue(2, 0, 0n).toString(), '2 months');
-    assert.equal(intervalValue(-2, 0, 0n).toString(), '-2 months');
-    assert.equal(intervalValue(12, 0, 0n).toString(), '1 year');
-    assert.equal(intervalValue(-12, 0, 0n).toString(), '-1 year');
-    assert.equal(intervalValue(24, 0, 0n).toString(), '2 years');
-    assert.equal(intervalValue(-24, 0, 0n).toString(), '-2 years');
-    assert.equal(intervalValue(25, 0, 0n).toString(), '2 years 1 month');
-    assert.equal(intervalValue(-25, 0, 0n).toString(), '-2 years -1 month');
-
-    assert.equal(intervalValue(0, 1, 0n).toString(), '1 day');
-    assert.equal(intervalValue(0, -1, 0n).toString(), '-1 day');
-    assert.equal(intervalValue(0, 2, 0n).toString(), '2 days');
-    assert.equal(intervalValue(0, -2, 0n).toString(), '-2 days');
-    assert.equal(intervalValue(0, 30, 0n).toString(), '30 days');
-    assert.equal(intervalValue(0, 365, 0n).toString(), '365 days');
-
-    assert.equal(intervalValue(0, 0, 1n).toString(), '00:00:00.000001');
-    assert.equal(intervalValue(0, 0, -1n).toString(), '-00:00:00.000001');
-    assert.equal(intervalValue(0, 0, 987654n).toString(), '00:00:00.987654');
-    assert.equal(intervalValue(0, 0, -987654n).toString(), '-00:00:00.987654');
-    assert.equal(intervalValue(0, 0, 1000000n).toString(), '00:00:01');
-    assert.equal(intervalValue(0, 0, -1000000n).toString(), '-00:00:01');
-    assert.equal(intervalValue(0, 0, 59n * 1000000n).toString(), '00:00:59');
-    assert.equal(intervalValue(0, 0, -59n * 1000000n).toString(), '-00:00:59');
-    assert.equal(intervalValue(0, 0, 60n * 1000000n).toString(), '00:01:00');
-    assert.equal(intervalValue(0, 0, -60n * 1000000n).toString(), '-00:01:00');
-    assert.equal(
-      intervalValue(0, 0, 59n * 60n * 1000000n).toString(),
-      '00:59:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, -59n * 60n * 1000000n).toString(),
-      '-00:59:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, 60n * 60n * 1000000n).toString(),
-      '01:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, -60n * 60n * 1000000n).toString(),
-      '-01:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, 24n * 60n * 60n * 1000000n).toString(),
-      '24:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, -24n * 60n * 60n * 1000000n).toString(),
-      '-24:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, 2147483647n * 60n * 60n * 1000000n).toString(),
-      '2147483647:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, -2147483647n * 60n * 60n * 1000000n).toString(),
-      '-2147483647:00:00',
-    );
-    assert.equal(
-      intervalValue(0, 0, 2147483647n * 60n * 60n * 1000000n + 1n).toString(),
-      '2147483647:00:00.000001',
-    );
-    assert.equal(
-      intervalValue(
-        0,
-        0,
-        -(2147483647n * 60n * 60n * 1000000n + 1n),
-      ).toString(),
-      '-2147483647:00:00.000001',
-    );
-
-    assert.equal(
-      intervalValue(
-        2 * 12 + 3,
-        5,
-        (7n * 60n * 60n + 11n * 60n + 13n) * 1000000n + 17n,
-      ).toString(),
-      '2 years 3 months 5 days 07:11:13.000017',
-    );
-    assert.equal(
-      intervalValue(
-        -(2 * 12 + 3),
-        -5,
-        -((7n * 60n * 60n + 11n * 60n + 13n) * 1000000n + 17n),
-      ).toString(),
-      '-2 years -3 months -5 days -07:11:13.000017',
-    );
-
-    // list
-    assert.equal(listValue([]).toString(), '[]');
-    assert.equal(listValue([1, 2, 3]).toString(), '[1, 2, 3]');
-    assert.equal(listValue(['a', 'b', 'c']).toString(), `['a', 'b', 'c']`);
-
-    // map
-    assert.equal(mapValue([]).toString(), '{}');
-    assert.equal(
-      mapValue([
-        { key: 1, value: 'a' },
-        { key: 2, value: 'b' },
-      ]).toString(),
-      `{1: 'a', 2: 'b'}`,
-    );
-
-    // struct
-    assert.equal(structValue({}).toString(), '{}');
-    assert.equal(structValue({ a: 1, b: 2 }).toString(), `{'a': 1, 'b': 2}`);
-
-    // timestamp milliseconds
-    assert.equal(TIMESTAMP_MS.epoch.toString(), '1970-01-01 00:00:00');
-    assert.equal(TIMESTAMP_MS.max.toString(), '294247-01-10 04:00:54.775');
-    assert.equal(TIMESTAMP_MS.min.toString(), '290309-12-22 (BC) 00:00:00');
-
-    // timestamp nanoseconds
-    assert.equal(TIMESTAMP_NS.epoch.toString(), '1970-01-01 00:00:00');
-    assert.equal(TIMESTAMP_NS.max.toString(), '2262-04-11 23:47:16.854775806');
-    assert.equal(TIMESTAMP_NS.min.toString(), '1677-09-22 00:00:00');
-
-    // timestamp seconds
-    assert.equal(TIMESTAMP_S.epoch.toString(), '1970-01-01 00:00:00');
-    assert.equal(TIMESTAMP_S.max.toString(), '294247-01-10 04:00:54');
-    assert.equal(TIMESTAMP_S.min.toString(), '290309-12-22 (BC) 00:00:00');
-
-    // timestamp tz
-    assert.equal(TIMESTAMPTZ.epoch.toString(), '1969-12-31 18:30:00-05:30');
-    assert.equal(
-      TIMESTAMPTZ.max.toString(),
-      '294247-01-09 22:30:54.775806-05:30',
-    );
-    assert.equal(
-      TIMESTAMPTZ.min.toString(),
-      '290309-12-21 (BC) 18:30:00-05:30',
-    );
-    assert.equal(TIMESTAMPTZ.posInf.toString(), 'infinity');
-    assert.equal(TIMESTAMPTZ.negInf.toString(), '-infinity');
-
-    // timestamp
-    assert.equal(TIMESTAMP.epoch.toString(), '1970-01-01 00:00:00');
-    assert.equal(TIMESTAMP.max.toString(), '294247-01-10 04:00:54.775806');
-    assert.equal(TIMESTAMP.min.toString(), '290309-12-22 (BC) 00:00:00');
-    assert.equal(TIMESTAMP.posInf.toString(), 'infinity');
-    assert.equal(TIMESTAMP.negInf.toString(), '-infinity');
-
-    // time tz
-    assert.equal(timeTZValue(0n, 0).toString(), '00:00:00+00');
-    assert.equal(
-      timeTZValue(
-        (((12n * 60n + 34n) * 60n + 56n) * 1000n + 789n) * 1000n,
-        -((7 * 60 + 9) * 60),
-      ).toString(),
-      '12:34:56.789-07:09',
-    );
-    assert.equal(TIMETZ.max.toString(), '24:00:00-15:59:59');
-    assert.equal(TIMETZ.min.toString(), '00:00:00+15:59:59');
-
-    // time
-    assert.equal(TIME.max.toString(), '24:00:00');
-    assert.equal(TIME.min.toString(), '00:00:00');
-    assert.equal(
-      timeValue(
-        (12n * 60n * 60n + 34n * 60n + 56n) * 1000000n + 987654n,
-      ).toString(),
-      '12:34:56.987654',
-    );
-
-    // union
-    assert.equal(unionValue('a', 42).toString(), '42');
-    assert.equal(unionValue('b', 'duck').toString(), 'duck');
-
-    // uuid
-    assert.equal(UUID.min.toString(), '00000000-0000-0000-0000-000000000000');
-    assert.equal(UUID.max.toString(), 'ffffffff-ffff-ffff-ffff-ffffffffffff');
-  });
-  test('date isFinite', () => {
-    assert.isTrue(DATE.epoch.isFinite);
-    assert.isTrue(DATE.max.isFinite);
-    assert.isTrue(DATE.min.isFinite);
-    assert.isFalse(DATE.posInf.isFinite);
-    assert.isFalse(DATE.negInf.isFinite);
-  });
-  test('timestamp isFinite', () => {
-    assert.isTrue(TIMESTAMP.epoch.isFinite);
-    assert.isTrue(TIMESTAMP.max.isFinite);
-    assert.isTrue(TIMESTAMP.min.isFinite);
-    assert.isFalse(TIMESTAMP.posInf.isFinite);
-    assert.isFalse(TIMESTAMP.negInf.isFinite);
-  });
-  test('timestamp_s isFinite', () => {
-    assert.isTrue(TIMESTAMP_S.epoch.isFinite);
-    assert.isTrue(TIMESTAMP_S.max.isFinite);
-    assert.isTrue(TIMESTAMP_S.min.isFinite);
-    assert.isFalse(TIMESTAMP_S.posInf.isFinite);
-    assert.isFalse(TIMESTAMP_S.negInf.isFinite);
-  });
-  test('timestamp_ms isFinite', () => {
-    assert.isTrue(TIMESTAMP_MS.epoch.isFinite);
-    assert.isTrue(TIMESTAMP_MS.max.isFinite);
-    assert.isTrue(TIMESTAMP_MS.min.isFinite);
-    assert.isFalse(TIMESTAMP_MS.posInf.isFinite);
-    assert.isFalse(TIMESTAMP_MS.negInf.isFinite);
-  });
-  test('timestamp_ns isFinite', () => {
-    assert.isTrue(TIMESTAMP_NS.epoch.isFinite);
-    assert.isTrue(TIMESTAMP_NS.max.isFinite);
-    assert.isTrue(TIMESTAMP_NS.min.isFinite);
-    assert.isFalse(TIMESTAMP_NS.posInf.isFinite);
-    assert.isFalse(TIMESTAMP_NS.negInf.isFinite);
-  });
-  test('value conversion', () => {
-    const dateParts: DateParts = { year: 2024, month: 6, day: 3 };
-    const timeParts: TimeParts = { hour: 12, min: 34, sec: 56, micros: 789123 };
-    const timeTZParts: TimeTZParts = {
-      time: timeParts,
-      offset: DuckDBTimeTZValue.MinOffset,
-    };
-    const timestampParts: TimestampParts = { date: dateParts, time: timeParts };
-
-    assert.deepEqual(DuckDBDateValue.fromParts(dateParts).toParts(), dateParts);
-    assert.deepEqual(DuckDBTimeValue.fromParts(timeParts).toParts(), timeParts);
-    assert.deepEqual(
-      DuckDBTimeTZValue.fromParts(timeTZParts).toParts(),
-      timeTZParts,
-    );
-    assert.deepEqual(
-      DuckDBTimestampValue.fromParts(timestampParts).toParts(),
-      timestampParts,
-    );
-    assert.deepEqual(
-      DuckDBTimestampTZValue.fromParts(timestampParts).toParts(),
-      timestampParts,
-    );
-
-    assert.deepEqual(
-      DuckDBDecimalValue.fromDouble(3.14159, 6, 5),
-      decimalValue(314159n, 6, 5),
-    );
-    assert.deepEqual(decimalValue(314159n, 6, 5).toDouble(), 3.14159);
-  });
-  test('result inspection conveniences', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.run(
-        'select i::int as a, i::int + 10 as b from range(3) t(i)',
-      );
-      assert.deepEqual(result.columnNames(), ['a', 'b']);
-      assert.deepEqual(result.columnTypes(), [INTEGER, INTEGER]);
-      const chunks = await result.fetchAllChunks();
-      const chunkColumns = chunks.map((chunk) => chunk.getColumns());
-      assert.deepEqual(chunkColumns, [
-        [
-          [0, 1, 2],
-          [10, 11, 12],
-        ],
-      ]);
-      const chunkRows = chunks.map((chunk) => chunk.getRows());
-      assert.deepEqual(chunkRows, [
-        [
-          [0, 10],
-          [1, 11],
-          [2, 12],
-        ],
-      ]);
-    });
-  });
-  test('row and column objects', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        'select i::int as a, i::int + 10 as b, (i + 100)::varchar as a from range(3) t(i)',
-      );
-      assert.deepEqual(reader.columnNames(), ['a', 'b', 'a']);
-      assert.deepEqual(reader.deduplicatedColumnNames(), ['a', 'b', 'a:1']);
-      assert.deepEqual(reader.columnTypes(), [INTEGER, INTEGER, VARCHAR]);
-      assert.deepEqual(reader.getRowObjects(), [
-        { 'a': 0, 'b': 10, 'a:1': '100' },
-        { 'a': 1, 'b': 11, 'a:1': '101' },
-        { 'a': 2, 'b': 12, 'a:1': '102' },
-      ]);
-      assert.deepEqual(reader.getColumnsObject(), {
-        'a': [0, 1, 2],
-        'b': [10, 11, 12],
-        'a:1': ['100', '101', '102'],
-      });
-    });
-  });
-  test('columns js', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(createTestJSQuery());
-      const columnsJS = reader.getColumnsJS();
-      assert.deepEqual(columnsJS, createTestJSColumnsJS());
-    });
-  });
-  test('columns object js', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(createTestJSQuery());
-      const columnsJS = reader.getColumnsObjectJS();
-      assert.deepEqual(columnsJS, createTestJSColumnsObjectJS());
-    });
-  });
-  test('rows js', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(createTestJSQuery());
-      const columnsJS = reader.getRowsJS();
-      assert.deepEqual(columnsJS, createTestJSRowsJS());
-    });
-  });
-  test('row objects js', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(createTestJSQuery());
-      const columnsJS = reader.getRowObjectsJS();
-      assert.deepEqual(columnsJS, createTestJSRowObjectsJS());
-    });
-  });
-  test('columns json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types()`,
-      );
-      const columnsJson = reader.getColumnsJson();
-      assert.deepEqual(columnsJson, createTestAllTypesColumnsJson());
-    });
-  });
-  test('columns object json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types()`,
-      );
-      const columnsJson = reader.getColumnsObjectJson();
-      assert.deepEqual(columnsJson, createTestAllTypesColumnsObjectJson());
-    });
-  });
-  test('rows json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types()`,
-      );
-      const rowsJson = reader.getRowsJson();
-      assert.deepEqual(rowsJson, createTestAllTypesRowsJson());
-    });
-  });
-  test('row objects json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types()`,
-      );
-      const rowObjectsJson = reader.getRowObjectsJson();
-      assert.deepEqual(rowObjectsJson, createTestAllTypesRowObjectsJson());
-    });
-  });
-  test('column names and types json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types(use_large_enum=true)`,
-      );
-      const columnNamesAndTypesJson = reader.columnNamesAndTypesJson();
-      assert.deepEqual(
-        columnNamesAndTypesJson,
-        createTestAllTypesColumnNamesAndTypesJson(),
-      );
-    });
-  });
-  test('column name and type objects json', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        `from test_all_types(use_large_enum=true)`,
-      );
-      const columnNameAndTypeObjectsJson =
-        reader.columnNameAndTypeObjectsJson();
-      assert.deepEqual(
-        columnNameAndTypeObjectsJson,
-        createTestAllTypesColumnNameAndTypeObjectsJson(),
-      );
-    });
-  });
-  test('result reader', async () => {
-    await withConnection(async (connection) => {
-      const reader = await connection.runAndReadAll(
-        'select i::int as a, i::int + 10000 as b from range(5000) t(i)',
-      );
-      assert.deepEqual(reader.columnNames(), ['a', 'b']);
-      assert.deepEqual(reader.columnTypes(), [INTEGER, INTEGER]);
-      const columns = reader.getColumns();
-      assert.equal(columns.length, 2);
-      assert.equal(columns[0][0], 0);
-      assert.equal(columns[0][4999], 4999);
-      assert.equal(columns[1][0], 10000);
-      assert.equal(columns[1][4999], 14999);
-      const rows = reader.getRows();
-      assert.equal(rows.length, 5000);
-      assert.deepEqual(rows[0], [0, 10000]);
-      assert.deepEqual(rows[4999], [4999, 14999]);
-    });
-  });
-  test('default duckdb_api without explicit options', async () => {
-    const instance = await DuckDBInstance.create();
-    const connection = await instance.connect();
-    const result = await connection.run(
-      `select current_setting('duckdb_api') as duckdb_api`,
-    );
-    assertColumns(result, [{ name: 'duckdb_api', type: VARCHAR }]);
-    const chunk = await result.fetchChunk();
-    assert.isDefined(chunk);
-    if (chunk) {
-      assert.strictEqual(chunk.columnCount, 1);
-      assert.strictEqual(chunk.rowCount, 1);
-      assertValues<string, DuckDBVarCharVector>(chunk, 0, DuckDBVarCharVector, [
-        'node-neo-api',
-      ]);
-    }
-  });
-  test('default duckdb_api with explicit options', async () => {
-    const instance = await DuckDBInstance.create(undefined, {});
-    const connection = await instance.connect();
-    const result = await connection.run(
-      `select current_setting('duckdb_api') as duckdb_api`,
-    );
-    assertColumns(result, [{ name: 'duckdb_api', type: VARCHAR }]);
-    const chunk = await result.fetchChunk();
-    assert.isDefined(chunk);
-    if (chunk) {
-      assert.strictEqual(chunk.columnCount, 1);
-      assert.strictEqual(chunk.rowCount, 1);
-      assertValues<string, DuckDBVarCharVector>(chunk, 0, DuckDBVarCharVector, [
-        'node-neo-api',
-      ]);
-    }
-  });
-  test('overriding duckdb_api', async () => {
-    const instance = await DuckDBInstance.create(undefined, {
-      'duckdb_api': 'custom-duckdb-api',
-    });
-    const connection = await instance.connect();
-    const result = await connection.run(
-      `select current_setting('duckdb_api') as duckdb_api`,
-    );
-    assertColumns(result, [{ name: 'duckdb_api', type: VARCHAR }]);
-    const chunk = await result.fetchChunk();
-    assert.isDefined(chunk);
-    if (chunk) {
-      assert.strictEqual(chunk.columnCount, 1);
-      assert.strictEqual(chunk.rowCount, 1);
-      assertValues<string, DuckDBVarCharVector>(chunk, 0, DuckDBVarCharVector, [
-        'custom-duckdb-api',
-      ]);
-    }
-  });
-  test('instance cache - same instance', async () => {
-    const cache = new DuckDBInstanceCache();
-    const instance1 = await cache.getOrCreateInstance(':memory:m1');
-    const connection1 = await instance1.connect();
-    await connection1.run(`create table t1 as select 1`);
-
-    const instance2 = await cache.getOrCreateInstance(':memory:m1');
-    const connection2 = await instance2.connect();
-    await connection2.run(`from t1`);
-  });
-  // Need to support explicitly destroying instance cache?
-  test.skip('instance cache - different instances', async () => {
-    let instance1: DuckDBInstance | undefined;
-    let instance2: DuckDBInstance | undefined;
-    try {
-      const cache = new DuckDBInstanceCache();
-      const instance1 = await cache.getOrCreateInstance(
-        'instance_cache_test_a.db',
-      );
-      const connection1 = await instance1.connect();
-      await connection1.run(`attach ':memory:' as mem1`);
-
-      const instance2 = await cache.getOrCreateInstance(
-        'instance_cache_test_b.db',
-      );
-      const connection2 = await instance2.connect();
-      try {
-        await connection2.run(`create table mem1.main.t1 as select 1`);
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(
-          err,
-          new Error(`Catalog Error: Catalog with name mem1 does not exist!`),
-        );
-      }
-    } finally {
-      if (instance1) {
-        instance1.closeSync();
-        fs.rmSync('instance_cache_test_a.db');
-      }
-      if (instance2) {
-        instance2.closeSync();
-        fs.rmSync('instance_cache_test_b.db');
-      }
-    }
-  });
-  test('instance cache - different config', async () => {
-    const cache = new DuckDBInstanceCache();
-    const instance1 = await cache.getOrCreateInstance(':memory:m1');
-    const connection1 = await instance1.connect();
-    await connection1.run(`create table t1 as select 1`);
-    try {
-      await cache.getOrCreateInstance(':memory:m1', {
-        accces_mode: 'READ_ONLY',
-      });
-      assert.fail('should throw');
-    } catch (err) {
-      assert.deepEqual(
-        err,
-        new Error(
-          `Connection Error: Can't open a connection to same database file with a different configuration than existing connections`,
-        ),
-      );
-    }
-  });
-  test('instance cache - singleton', async () => {
-    const instance = await DuckDBInstance.fromCache();
-    const connection = await instance.connect();
-    await connection.run('select 1');
-  });
-  test('create connection using instance cache', async () => {
-    const connection = await DuckDBConnection.create();
-    await connection.run('select 1');
-  });
-  test('data chunk max rows', () => {
-    try {
-      DuckDBDataChunk.create([INTEGER], 2049);
-      assert.fail('should throw');
-    } catch (err) {
-      assert.deepEqual(
-        err,
-        new Error('A data chunk cannot have more than 2048 rows'),
-      );
-    }
-  });
-  test('write integer vector', () => {
-    const chunk = DuckDBDataChunk.create([INTEGER], 3);
-    const vector = chunk.getColumnVector(0) as DuckDBIntegerVector;
-    assert.equal(vector.itemCount, 3);
-    vector.setItem(0, 42);
-    vector.setItem(1, 12345);
-    vector.setItem(2, 67890);
-    assert.equal(vector.getItem(0), 42);
-    assert.equal(vector.getItem(1), 12345);
-    assert.equal(vector.getItem(2), 67890);
-  });
-  test('write integer vector with nulls', () => {
-    const chunk = DuckDBDataChunk.create([INTEGER], 3);
-    const vector = chunk.getColumnVector(0) as DuckDBIntegerVector;
-    assert.equal(vector.itemCount, 3);
-    vector.setItem(0, 42);
-    vector.setItem(1, 12345);
-    vector.setItem(2, null);
-    assert.equal(vector.getItem(0), 42);
-    assert.equal(vector.getItem(1), 12345);
-    assert.equal(vector.getItem(2), null);
-  });
-  test('write list vector', () => {
-    const chunk = DuckDBDataChunk.create([LIST(INTEGER)], 3);
-    const vector = chunk.getColumnVector(0) as DuckDBListVector;
-    assert.equal(vector.itemCount, 3);
-    vector.setItem(0, listValue([10, 11, 12]));
-    vector.setItem(1, listValue([20, 21, 22]));
-    vector.setItem(2, null);
-    assert.deepEqual(vector.getItem(0), listValue([10, 11, 12]));
-    assert.deepEqual(vector.getItem(1), listValue([20, 21, 22]));
-    assert.equal(vector.getItem(2), null);
-  });
-  test('create and append data chunk', async () => {
-    await withConnection(async (connection) => {
-      const values = [42, 12345, null];
-
-      const chunk = DuckDBDataChunk.create([INTEGER], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 int)');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBIntegerVector, values);
-      }
-    });
-  });
-  test('create and append data chunk, delayed flush', async () => {
-    await withConnection(async (connection) => {
-      const baselineValues = [10, 11, 12];
-      const targetValues = [42, 12345, null];
-
-      const chunk = DuckDBDataChunk.create([INTEGER], targetValues.length);
-      const vector = chunk.getColumnVector(0) as DuckDBIntegerVector;
-
-      // First, set the values to something known as a baseline.
-      for (let i = 0; i < baselineValues.length; i++) {
-        vector.setItem(i, baselineValues[i]);
-      }
-      vector.flush();
-
-      // Then, set the values to our target, but don't flush (yet).
-      for (let i = 0; i < targetValues.length; i++) {
-        vector.setItem(i, targetValues[i]);
-      }
-
-      await connection.run('create table target1(col0 int)');
-      const appender1 = await connection.createAppender('target1');
-      appender1.appendDataChunk(chunk);
-      appender1.flushSync();
-
-      const result1 = await connection.run('from target1');
-      const result1Chunk = await result1.fetchChunk();
-      assert.isDefined(result1Chunk);
-      if (result1Chunk) {
-        assert.equal(result1Chunk.columnCount, 1);
-        assert.equal(result1Chunk.rowCount, targetValues.length);
-        assertValues(result1Chunk, 0, DuckDBIntegerVector, baselineValues);
-      }
-
-      // now flush
-      vector.flush();
-
-      await connection.run('create table target2(col0 int)');
-      const appender2 = await connection.createAppender('target2');
-      appender2.appendDataChunk(chunk);
-      appender2.flushSync();
-
-      const result2 = await connection.run('from target2');
-      const result2Chunk = await result2.fetchChunk();
-      assert.isDefined(result2Chunk);
-      if (result2Chunk) {
-        assert.equal(result2Chunk.columnCount, 1);
-        assert.equal(result2Chunk.rowCount, 3);
-        assertValues(result2Chunk, 0, DuckDBIntegerVector, targetValues);
-      }
-    });
-  });
-  test('create and append data chunk with varchars', async () => {
-    await withConnection(async (connection) => {
-      const values = ['xyz', 'abcdefghijkl', 'ABCDEFGHIJKLM', null];
-
-      const chunk = DuckDBDataChunk.create([VARCHAR], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 varchar)');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBVarCharVector, values);
-      }
-    });
-  });
-  test('create and append data chunk with blobs', async () => {
-    await withConnection(async (connection) => {
-      const values = [
-        blobValue(new Uint8Array([0xab, 0xcd, 0xef])),
-        blobValue(
-          new Uint8Array([
-            0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b,
-            0x0c,
-          ]),
-        ),
-        blobValue(
-          new Uint8Array([
-            0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b,
-            0x1c, 0x1d,
-          ]),
-        ),
-        null,
-      ];
-      const chunk = DuckDBDataChunk.create([BLOB], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 blob)');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBBlobVector, values);
-      }
-    });
-  });
-  test('create and append data chunk with lists', async () => {
-    await withConnection(async (connection) => {
-      const values = [
-        listValue([10, 11, 12]),
-        listValue([]),
-        listValue([100, 200]),
-        null,
-      ];
-
-      const chunk = DuckDBDataChunk.create([LIST(INTEGER)], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 integer[])');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBListVector, values);
-      }
-    });
-  });
-  test('create and append data chunk, modify nested list vector', async () => {
-    await withConnection(async (connection) => {
-      const originalValues = [
-        listValue([listValue([110, 111]), listValue([]), listValue([130])]),
-        listValue([]),
-        listValue([
-          listValue([310, 311, 312]),
-          listValue([320, 321]),
-          listValue([330, 331, 332, 333]),
-        ]),
-      ];
-
-      const chunk = DuckDBDataChunk.create(
-        [LIST(LIST(INTEGER))],
-        originalValues.length,
-      );
-      chunk.setColumnValues(0, originalValues);
-
-      const outerListVector = chunk.getColumnVector(0) as DuckDBListVector;
-      const innerListVector = outerListVector.getItemVector(
-        2,
-      ) as DuckDBListVector;
-      innerListVector.setItem(1, listValue([350, 351, 352, 353, 354]));
-      innerListVector.flush();
-
-      const modifiedValues = [...originalValues];
-      modifiedValues[2] = listValue([
-        listValue([310, 311, 312]),
-        listValue([350, 351, 352, 353, 354]),
-        listValue([330, 331, 332, 333]),
-      ]);
-
-      await connection.run('create table target(col0 integer[][])');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, modifiedValues.length);
-        assertValues(resultChunk, 0, DuckDBListVector, modifiedValues);
-      }
-    });
-  });
-  test('create and append data chunk with arrays of integers', async () => {
-    await withConnection(async (connection) => {
-      const values = [
-        arrayValue([10, 11, 12]),
-        arrayValue([20, 21, 22]),
-        arrayValue([30, 31, 32]),
-        null,
-      ];
-
-      const chunk = DuckDBDataChunk.create([ARRAY(INTEGER, 3)], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 integer[3])');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBArrayVector, values);
-      }
-    });
-  });
-  test('create and append data chunk with arrays of varchars', async () => {
-    await withConnection(async (connection) => {
-      const values = [
-        arrayValue(['a', 'b', 'c']),
-        arrayValue(['d', 'e', 'f']),
-        arrayValue(['g', 'h', 'i']),
-        null,
-      ];
-
-      const chunk = DuckDBDataChunk.create([ARRAY(VARCHAR, 3)], values.length);
-      chunk.setColumnValues(0, values);
-
-      await connection.run('create table target(col0 varchar[3])');
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBArrayVector, values);
-      }
-    });
-  });
-  test('create and append data chunk with structs', async () => {
-    await withConnection(async (connection) => {
-      const values = [
-        structValue({ 'num': 10, 'str': 'xyz' }),
-        structValue({ 'num': 11, 'str': 'abcdefghijkl' }),
-        structValue({ 'num': 12, 'str': 'ABCDEFGHIJKLM' }),
-        null,
-      ];
-
-      const chunk = DuckDBDataChunk.create(
-        [STRUCT({ 'num': INTEGER, 'str': VARCHAR })],
-        values.length,
-      );
-      chunk.setColumnValues(0, values);
-
-      await connection.run(
-        'create table target(col0 struct(num integer, str varchar))',
-      );
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.equal(resultChunk.columnCount, 1);
-        assert.equal(resultChunk.rowCount, values.length);
-        assertValues(resultChunk, 0, DuckDBStructVector, values);
-      }
-    });
-  });
-  test('create and append data chunk with rows', async () => {
-    await withConnection(async (connection) => {
-      const types: DuckDBType[] = [BOOLEAN, TINYINT, SMALLINT, INTEGER];
-      const columnData: DuckDBValue[][] = [
-        [false, true, null],
-        [TINYINT.min, TINYINT.max, null],
-        [SMALLINT.min, SMALLINT.max, null],
-        [INTEGER.min, INTEGER.max, null],
-      ];
-      const rows: DuckDBValue[][] = [
-        [false, TINYINT.min, SMALLINT.min, INTEGER.min],
-        [true, TINYINT.max, SMALLINT.max, INTEGER.max],
-        [null, null, null, null],
-      ];
-
-      const chunk = DuckDBDataChunk.create(types);
-      chunk.setRows(rows);
-
-      await connection.run(
-        'create table target(\
-          bool boolean,\
-          tinyint tinyint,\
-          smallint smallint,\
-          int integer\
-        )',
-      );
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.strictEqual(resultChunk.columnCount, types.length);
-        assert.strictEqual(resultChunk.rowCount, 3);
-        assertValues(resultChunk, 0, DuckDBBooleanVector, columnData[0]);
-        assertValues(resultChunk, 1, DuckDBTinyIntVector, columnData[1]);
-        assertValues(resultChunk, 2, DuckDBSmallIntVector, columnData[2]);
-        assertValues(resultChunk, 3, DuckDBIntegerVector, columnData[3]);
-      }
-    });
-  });
-  test('create and append data chunk with all types', async () => {
-    await withConnection(async (connection) => {
-      const types = createTestAllTypesColumnTypes();
-      const columns = createTestAllTypesColumns();
-      const columnNameAndTypeObjects =
-        createTestAllTypesColumnNameAndTypeObjects();
-
-      const chunk = DuckDBDataChunk.create(types);
-      chunk.setColumns(columns);
-
-      await connection.run(
-        `create table target(${columnNameAndTypeObjects
-          .map(({ name, type }) => `"${name.replace(`"`, `""`)}" ${type}`)
-          .join(', ')})`,
-      );
-      const appender = await connection.createAppender('target');
-      appender.appendDataChunk(chunk);
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.strictEqual(resultChunk.columnCount, columns.length);
-        assert.strictEqual(resultChunk.rowCount, 3);
-        assertValues(resultChunk, 0, DuckDBBooleanVector, columns[0]);
-        assertValues(resultChunk, 1, DuckDBTinyIntVector, columns[1]);
-        assertValues(resultChunk, 2, DuckDBSmallIntVector, columns[2]);
-        assertValues(resultChunk, 3, DuckDBIntegerVector, columns[3]);
-        assertValues(resultChunk, 4, DuckDBBigIntVector, columns[4]);
-        assertValues(resultChunk, 5, DuckDBHugeIntVector, columns[5]);
-        assertValues(resultChunk, 6, DuckDBUHugeIntVector, columns[6]);
-        assertValues(resultChunk, 7, DuckDBUTinyIntVector, columns[7]);
-        assertValues(resultChunk, 8, DuckDBUSmallIntVector, columns[8]);
-        assertValues(resultChunk, 9, DuckDBUIntegerVector, columns[9]);
-        assertValues(resultChunk, 10, DuckDBUBigIntVector, columns[10]);
-        assertValues(resultChunk, 11, DuckDBBigNumVector, columns[11]);
-        assertValues(resultChunk, 12, DuckDBDateVector, columns[12]);
-        assertValues(resultChunk, 13, DuckDBTimeVector, columns[13]);
-        assertValues(resultChunk, 14, DuckDBTimestampVector, columns[14]);
-        assertValues(
-          resultChunk,
-          15,
-          DuckDBTimestampSecondsVector,
-          columns[15],
-        );
-        assertValues(
-          resultChunk,
-          16,
-          DuckDBTimestampMillisecondsVector,
-          columns[16],
-        );
-        assertValues(
-          resultChunk,
-          17,
-          DuckDBTimestampNanosecondsVector,
-          columns[17],
-        );
-        assertValues(resultChunk, 18, DuckDBTimeTZVector, columns[18]);
-        assertValues(resultChunk, 19, DuckDBTimestampTZVector, columns[19]);
-        assertValues(resultChunk, 20, DuckDBFloatVector, columns[20]);
-        assertValues(resultChunk, 21, DuckDBDoubleVector, columns[21]);
-        assertValues(resultChunk, 22, DuckDBDecimal16Vector, columns[22]);
-        assertValues(resultChunk, 23, DuckDBDecimal32Vector, columns[23]);
-        assertValues(resultChunk, 24, DuckDBDecimal64Vector, columns[24]);
-        assertValues(resultChunk, 25, DuckDBDecimal128Vector, columns[25]);
-        assertValues(resultChunk, 26, DuckDBUUIDVector, columns[26]);
-        assertValues(resultChunk, 27, DuckDBIntervalVector, columns[27]);
-        assertValues(resultChunk, 28, DuckDBVarCharVector, columns[28]);
-        assertValues(resultChunk, 29, DuckDBBlobVector, columns[29]);
-        assertValues(resultChunk, 30, DuckDBBitVector, columns[30]);
-        assertValues(resultChunk, 31, DuckDBEnum8Vector, columns[31]);
-        assertValues(resultChunk, 32, DuckDBEnum16Vector, columns[32]);
-        assertValues(resultChunk, 33, DuckDBEnum32Vector, columns[33]);
-        assertValues(resultChunk, 34, DuckDBListVector, columns[34]); // int_array
-        assertValues(resultChunk, 35, DuckDBListVector, columns[35]); // double_array
-        assertValues(resultChunk, 36, DuckDBListVector, columns[36]); // date_array
-        assertValues(resultChunk, 37, DuckDBListVector, columns[37]); // timestamp_array
-        assertValues(resultChunk, 38, DuckDBListVector, columns[38]); // timestamptz_array
-        assertValues(resultChunk, 39, DuckDBListVector, columns[39]); // varchar_array
-        assertValues(resultChunk, 40, DuckDBListVector, columns[40]); // nested_int_array
-        assertValues(resultChunk, 41, DuckDBStructVector, columns[41]);
-        assertValues(resultChunk, 42, DuckDBStructVector, columns[42]); // struct_of_arrays
-        assertValues(resultChunk, 43, DuckDBListVector, columns[43]); // array_of_structs
-        assertValues(resultChunk, 44, DuckDBMapVector, columns[44]);
-        assertValues(resultChunk, 45, DuckDBUnionVector, columns[45]);
-        assertValues(resultChunk, 46, DuckDBArrayVector, columns[46]); // fixed_int_array
-        assertValues(resultChunk, 47, DuckDBArrayVector, columns[47]); // fixed_varchar_array
-        assertValues(resultChunk, 48, DuckDBArrayVector, columns[48]); // fixed_nested_int_array
-        assertValues(resultChunk, 49, DuckDBArrayVector, columns[49]); // fixed_nested_varchar_array
-        assertValues(resultChunk, 50, DuckDBArrayVector, columns[50]); // fixed_struct_array
-        assertValues(resultChunk, 51, DuckDBStructVector, columns[51]); // struct_of_fixed_array
-        assertValues(resultChunk, 52, DuckDBArrayVector, columns[52]); // fixed_array_of_int_list
-        assertValues(resultChunk, 53, DuckDBListVector, columns[53]); // list_of_fixed_int_array
-      }
-    });
-  });
-  test('append all types row-by-row', async () => {
-    await withConnection(async (connection) => {
-      const types = createTestAllTypesColumnTypes();
-      const columns = createTestAllTypesColumns();
-      const columnNameAndTypeObjects =
-        createTestAllTypesColumnNameAndTypeObjects();
-
-      await connection.run(
-        `create table target(${columnNameAndTypeObjects
-          .map(({ name, type }) => `"${name.replace(`"`, `""`)}" ${type}`)
-          .join(', ')})`,
-      );
-
-      const appender = await connection.createAppender('target');
-      for (let rowIndex = 0; rowIndex < columns[0].length; rowIndex++) {
-        for (let columnIndex = 0; columnIndex < columns.length; columnIndex++) {
-          const type = types[columnIndex];
-          const value = columns[columnIndex][rowIndex];
-          if (value === null) {
-            appender.appendNull();
-          } else {
-            if (type.typeId === DuckDBTypeId.GEOMETRY) {
-              // TODO: Appending GEOMETRY is not yet supported
-              appender.appendNull();
-            } else {
-              appender.appendValue(value, type);
-            }
-          }
-        }
-        appender.endRow();
-      }
-      appender.flushSync();
-
-      const result = await connection.run('from target');
-      const resultChunk = await result.fetchChunk();
-      assert.isDefined(resultChunk);
-      if (resultChunk) {
-        assert.strictEqual(resultChunk.columnCount, columns.length);
-        assert.strictEqual(resultChunk.rowCount, 3);
-        assertValues(resultChunk, 0, DuckDBBooleanVector, columns[0]);
-        assertValues(resultChunk, 1, DuckDBTinyIntVector, columns[1]);
-        assertValues(resultChunk, 2, DuckDBSmallIntVector, columns[2]);
-        assertValues(resultChunk, 3, DuckDBIntegerVector, columns[3]);
-        assertValues(resultChunk, 4, DuckDBBigIntVector, columns[4]);
-        assertValues(resultChunk, 5, DuckDBHugeIntVector, columns[5]);
-        assertValues(resultChunk, 6, DuckDBUHugeIntVector, columns[6]);
-        assertValues(resultChunk, 7, DuckDBUTinyIntVector, columns[7]);
-        assertValues(resultChunk, 8, DuckDBUSmallIntVector, columns[8]);
-        assertValues(resultChunk, 9, DuckDBUIntegerVector, columns[9]);
-        assertValues(resultChunk, 10, DuckDBUBigIntVector, columns[10]);
-        assertValues(resultChunk, 11, DuckDBBigNumVector, columns[11]);
-        assertValues(resultChunk, 12, DuckDBDateVector, columns[12]);
-        assertValues(resultChunk, 13, DuckDBTimeVector, columns[13]);
-        assertValues(resultChunk, 14, DuckDBTimestampVector, columns[14]);
-        assertValues(
-          resultChunk,
-          15,
-          DuckDBTimestampSecondsVector,
-          columns[15],
-        );
-        assertValues(
-          resultChunk,
-          16,
-          DuckDBTimestampMillisecondsVector,
-          columns[16],
-        );
-        assertValues(
-          resultChunk,
-          17,
-          DuckDBTimestampNanosecondsVector,
-          columns[17],
-        );
-        assertValues(resultChunk, 18, DuckDBTimeTZVector, columns[18]);
-        assertValues(resultChunk, 19, DuckDBTimestampTZVector, columns[19]);
-        assertValues(resultChunk, 20, DuckDBFloatVector, columns[20]);
-        assertValues(resultChunk, 21, DuckDBDoubleVector, columns[21]);
-        assertValues(resultChunk, 22, DuckDBDecimal16Vector, columns[22]);
-        assertValues(resultChunk, 23, DuckDBDecimal32Vector, columns[23]);
-        assertValues(resultChunk, 24, DuckDBDecimal64Vector, columns[24]);
-        assertValues(resultChunk, 25, DuckDBDecimal128Vector, columns[25]);
-        assertValues(resultChunk, 26, DuckDBUUIDVector, columns[26]);
-        assertValues(resultChunk, 27, DuckDBIntervalVector, columns[27]);
-        assertValues(resultChunk, 28, DuckDBVarCharVector, columns[28]);
-        assertValues(resultChunk, 29, DuckDBBlobVector, columns[29]);
-        assertValues(resultChunk, 30, DuckDBBitVector, columns[30]);
-        assertValues(resultChunk, 31, DuckDBEnum8Vector, columns[31]);
-        assertValues(resultChunk, 32, DuckDBEnum16Vector, columns[32]);
-        assertValues(resultChunk, 33, DuckDBEnum32Vector, columns[33]);
-        assertValues(resultChunk, 34, DuckDBListVector, columns[34]); // int_array
-        assertValues(resultChunk, 35, DuckDBListVector, columns[35]); // double_array
-        assertValues(resultChunk, 36, DuckDBListVector, columns[36]); // date_array
-        assertValues(resultChunk, 37, DuckDBListVector, columns[37]); // timestamp_array
-        assertValues(resultChunk, 38, DuckDBListVector, columns[38]); // timestamptz_array
-        assertValues(resultChunk, 39, DuckDBListVector, columns[39]); // varchar_array
-        assertValues(resultChunk, 40, DuckDBListVector, columns[40]); // nested_int_array
-        assertValues(resultChunk, 41, DuckDBStructVector, columns[41]);
-        assertValues(resultChunk, 42, DuckDBStructVector, columns[42]); // struct_of_arrays
-        assertValues(resultChunk, 43, DuckDBListVector, columns[43]); // array_of_structs
-        assertValues(resultChunk, 44, DuckDBMapVector, columns[44]);
-        assertValues(resultChunk, 45, DuckDBUnionVector, columns[45]);
-        assertValues(resultChunk, 46, DuckDBArrayVector, columns[46]); // fixed_int_array
-        assertValues(resultChunk, 47, DuckDBArrayVector, columns[47]); // fixed_varchar_array
-        assertValues(resultChunk, 48, DuckDBArrayVector, columns[48]); // fixed_nested_int_array
-        assertValues(resultChunk, 49, DuckDBArrayVector, columns[49]); // fixed_nested_varchar_array
-        assertValues(resultChunk, 50, DuckDBArrayVector, columns[50]); // fixed_struct_array
-        assertValues(resultChunk, 51, DuckDBStructVector, columns[51]); // struct_of_fixed_array
-        assertValues(resultChunk, 52, DuckDBArrayVector, columns[52]); // fixed_array_of_int_list
-        assertValues(resultChunk, 53, DuckDBListVector, columns[53]); // list_of_fixed_int_array
-        assertValues(resultChunk, 54, DuckDBTimeNSVector, columns[54]); // time_ns
-      }
-    });
-  });
-
-  test('should not segfault with concurrent prepared statement creation and execution', async () => {
-    await withConnection(async (connection) => {
-      // Create test table with some data
-      await connection.run(`
-        CREATE TABLE test_table (
-          id INTEGER, 
-          name VARCHAR, 
-          value INTEGER, 
-          created_at VARCHAR
-        )
-      `);
-
-      await connection.run(`
-        INSERT INTO test_table VALUES 
-        (1, 'test1', 100, '2023-01-01'),
-        (2, 'test2', 200, '2023-01-02'),
-        (3, 'test3', 300, '2023-01-03'),
-        (4, 'test4', 400, '2023-01-04'),
-        (5, 'test5', 500, '2023-01-05')
-      `);
-
-      const iterations = 1000;
-      const concurrency = 12;
-
-      const runIteration = async (i: number) => {
-        const prepared = await connection.prepare(
-          'SELECT * FROM test_table WHERE id = $1',
-        );
-        prepared.bindInteger(1, (i % 5) + 1);
-
-        await prepared.run();
-      };
-
-      // Run iterations in batches with controlled concurrency
-      let processed = 0;
-
-      while (processed < iterations) {
-        const batch = [];
-        const batchEnd = Math.min(processed + concurrency, iterations);
-
-        for (let i = processed; i < batchEnd; i++) {
-          batch.push(runIteration(i));
-        }
-
-        // Wait for all in the batch to complete
-        await Promise.allSettled(batch);
-        processed = batchEnd;
-      }
-
-      // If we reach here without segfaulting, the test passes
-      assert.isTrue(true, 'Test completed without error');
-    });
-  });
-
-  test('scalar function (no params)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (_info, input, output) => {
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(rowIndex, `my_output_${rowIndex}`);
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      const reader = await connection.runAndReadAll('select my_func()');
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, { 'my_func()': ['my_output_0'] });
-    });
-  });
-
-  test('scalar function (bind data)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          bindFunction: (info) => {
-            info.setBindData({ 'my_bind_data_key': 'my_bind_data_value' });
-          },
-          mainFunction: (info, input, output) => {
-            const { bindData } = info;
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(
-                rowIndex,
-                `my_output_${rowIndex}_${JSON.stringify(bindData)}`,
-              );
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      const reader = await connection.runAndReadAll('select my_func()');
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func()': [
-          `my_output_0_${JSON.stringify({ 'my_bind_data_key': 'my_bind_data_value' })}`,
-        ],
-      });
-    });
-  });
-
-  test('scalar function (extra info)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (info, input, output) => {
-            const { extraInfo } = info;
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(
-                rowIndex,
-                `my_output_${rowIndex}_${JSON.stringify(extraInfo)}`,
-              );
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-          extraInfo: { 'my_extra_info_key': 'my_extra_info_value' },
-        }),
-      );
-      const reader = await connection.runAndReadAll('select my_func()');
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func()': [
-          'my_output_0_{"my_extra_info_key":"my_extra_info_value"}',
-        ],
-      });
-    });
-  });
-
-  test('scalar function (extra info & bind data)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          bindFunction: (info) => {
-            const { extraInfo, clientContext } = info;
-            info.setBindData({
-              'my_bind_data_key': 'my_bind_data_value',
-              'extra_info': extraInfo,
-              'valid_connection_id': clientContext.connectionId > 0,
-            });
-          },
-          mainFunction: (info, input, output) => {
-            const { bindData } = info;
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(
-                rowIndex,
-                `my_output_${rowIndex}_${JSON.stringify(bindData)}`,
-              );
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-          extraInfo: { 'my_extra_info_key': 'my_extra_info_value' },
-        }),
-      );
-      const reader = await connection.runAndReadAll('select my_func()');
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func()': [
-          `my_output_0_${JSON.stringify({
-            'my_bind_data_key': 'my_bind_data_value',
-            'extra_info': { 'my_extra_info_key': 'my_extra_info_value' },
-            'valid_connection_id': true,
-          })}`,
-        ],
-      });
-    });
-  });
-
-  test('scalar function (error handling: exception in main func)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (_info, _input, _output) => {
-            throw new Error('my_error');
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      try {
-        await connection.run('select my_func()');
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(err, new Error('Invalid Input Error: my_error'));
-      }
-    });
-  });
-
-  test('scalar function (error handling: setError in main func)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (info, _input, _output) => {
-            info.setError('my_error');
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      try {
-        await connection.run('select my_func()');
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(err, new Error('Invalid Input Error: my_error'));
-      }
-    });
-  });
-
-  test('scalar function (error handling: exception in bind func)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          bindFunction: (_info) => {
-            throw new Error('my_bind_error');
-          },
-          mainFunction: (_info, _input, _output) => {
-            throw new Error('my_error');
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      try {
-        await connection.run('select my_func()');
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(err, new Error('Binder Error: my_bind_error'));
-      }
-    });
-  });
-
-  test('scalar function (error handling: setError in bind func)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          bindFunction: (info) => {
-            info.setError('my_bind_error');
-          },
-          mainFunction: (info, _input, _output) => {
-            info.setError('my_error');
-          },
-          returnType: VARCHAR,
-        }),
-      );
-      try {
-        await connection.run('select my_func()');
-        assert.fail('should throw');
-      } catch (err) {
-        assert.deepEqual(err, new Error('Binder Error: my_bind_error'));
-      }
-    });
-  });
-
-  test('scalar function (params, volatile)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (_info, input, output) => {
-            const v0 = input.getColumnVector(0);
-            const v1 = input.getColumnVector(1);
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(
-                rowIndex,
-                `my_output_${rowIndex}_${v0.getItem(rowIndex)}_${v1.getItem(
-                  rowIndex,
-                )}`,
-              );
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-          parameterTypes: [INTEGER, VARCHAR],
-          volatile: true,
-        }),
-      );
-      const reader = await connection.runAndReadAll(
-        `select my_func(42, 'duck') as my_func_result from range(3)`,
-      );
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func_result': [
-          'my_output_0_42_duck',
-          'my_output_1_42_duck',
-          'my_output_2_42_duck',
-        ],
-      });
-    });
-  });
-
-  test('scalar function (varargs, volatile)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (_info, input, output) => {
-            // const v0 = input.getColumnVector(0);
-            // const v1 = input.getColumnVector(1);
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              const argValues: DuckDBValue[] = [];
-              for (
-                let columnIndex = 0;
-                columnIndex < input.columnCount;
-                columnIndex++
-              ) {
-                argValues.push(
-                  input.getColumnVector(columnIndex).getItem(rowIndex),
-                );
-              }
-              output.setItem(
-                rowIndex,
-                `my_output_${rowIndex}_${argValues.join('_')}`,
-              );
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-          varArgsType: INTEGER,
-          volatile: true,
-        }),
-      );
-      const reader = await connection.runAndReadAll(
-        `select my_func(11, 13, 17) as my_func_result from range(3)`,
-      );
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func_result': [
-          'my_output_0_11_13_17',
-          'my_output_1_11_13_17',
-          'my_output_2_11_13_17',
-        ],
-      });
-    });
-  });
-
-  test('scalar function (special handling)', async () => {
-    await withConnection(async (connection) => {
-      connection.registerScalarFunction(
-        DuckDBScalarFunction.create({
-          name: 'my_func',
-          mainFunction: (_info, input, output) => {
-            for (let rowIndex = 0; rowIndex < input.rowCount; rowIndex++) {
-              output.setItem(rowIndex, 'output_is_not_null');
-            }
-            output.flush();
-          },
-          returnType: VARCHAR,
-          parameterTypes: [INTEGER],
-          specialHandling: true,
-        }),
-      );
-      const reader = await connection.runAndReadAll(`select my_func(NULL)`);
-      const columns = reader.getColumnsObject();
-      assert.deepEqual(columns, {
-        'my_func(NULL)': ['output_is_not_null'],
-      });
-    });
-  });
-
-  describe('SQL utility functions', () => {
-    test('quotedString', () => {
-      // Basic string
-      assert.equal(quotedString('hello'), "'hello'");
-
-      // String with single quotes
-      assert.equal(quotedString("it's"), "'it''s'");
-      assert.equal(quotedString("'quoted'"), "'''quoted'''");
-
-      // String with multiple single quotes
-      assert.equal(
-        quotedString("it's 'really' good"),
-        "'it''s ''really'' good'",
-      );
-
-      // Empty string
-      assert.equal(quotedString(''), "''");
-
-      // String with special characters
-      assert.equal(quotedString('hello\nworld'), "'hello\nworld'");
-      assert.equal(quotedString('tab\there'), "'tab\there'");
-    });
-
-    test('quotedIdentifier', () => {
-      // Basic identifier
-      assert.equal(quotedIdentifier('table_name'), '"table_name"');
-
-      // Identifier with double quotes
-      assert.equal(quotedIdentifier('my"table'), '"my""table"');
-      assert.equal(quotedIdentifier('"column"'), '"""column"""');
-
-      // Identifier with multiple double quotes
-      assert.equal(
-        quotedIdentifier('my"special"table'),
-        '"my""special""table"',
-      );
-
-      // Empty identifier
-      assert.equal(quotedIdentifier(''), '""');
-
-      // Identifier with special characters
-      assert.equal(quotedIdentifier('table-name'), '"table-name"');
-      assert.equal(quotedIdentifier('table.name'), '"table.name"');
-      assert.equal(quotedIdentifier('table name'), '"table name"');
-    });
-  });
-
-  test('iterate over DuckDBResult stream in chunks', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(
-        'select i::int, i::int + 10, (i + 100)::varchar from range(3) t(i)',
-      );
-
-      for await (const chunk of result) {
-        assert.strictEqual(chunk.rowCount, 3);
-        let i = 0;
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          i++,
-          DuckDBIntegerVector,
-          [0, 1, 2],
-        );
-        assertValues<number, DuckDBIntegerVector>(
-          chunk,
-          i++,
-          DuckDBIntegerVector,
-          [10, 11, 12],
-        );
-        assertValues<string, DuckDBVarCharVector>(
-          chunk,
-          i++,
-          DuckDBVarCharVector,
-          ['100', '101', '102'],
-        );
-      }
-    });
-  });
-
-  test('iterate over many DuckDBResult chunks', async () => {
-    await withConnection(async (connection) => {
-      const chunkSize = 2048;
-      const totalExpectedCount = chunkSize * 3;
-      const result = await connection.stream(
-        `select i::int from range(${totalExpectedCount}) t(i)`,
-      );
-
-      let total = 0;
-      for await (const chunk of result) {
-        assert.equal(chunk.rowCount, chunkSize);
-        total += chunk.rowCount;
-      }
-
-      assert.equal(total, totalExpectedCount);
-    });
-  });
-
-  test('iterate stream of rows', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(
-        'select i::int, i::int + 10, (i + 100)::varchar from range(3) t(i)',
-      );
-
-      const expectedRows: DuckDBValue[][] = [
-        [0, 10, '100'],
-        [1, 11, '101'],
-        [2, 12, '102'],
-      ];
-
-      for await (const rows of result.yieldRows()) {
-        for (let i = 0; i < rows.length; i++) {
-          assert.deepEqual(rows[i], expectedRows[i]);
-        }
-      }
-    });
-  });
-
-  test('iterate stream of row objects', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(
-        'select i::int as a, i::int + 10 as b, (i + 100)::varchar as c from range(3) t(i)',
-      );
-
-      const expectedRows: Record<string, DuckDBValue>[] = [
-        { a: 0, b: 10, c: '100' },
-        { a: 1, b: 11, c: '101' },
-        { a: 2, b: 12, c: '102' },
-      ];
-
-      for await (const rows of result.yieldRowObjects()) {
-        for (let i = 0; i < rows.length; i++) {
-          assert.deepEqual(rows[i], expectedRows[i]);
-        }
-      }
-    });
-  });
-
-  test('iterate result stream rows js', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(createTestJSQuery());
-      for await (const row of result.yieldRowsJs()) {
-        assert.deepEqual(row, createTestJSRowsJS());
-      }
-    });
-  });
-
-  test('iterate result stream object js', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(createTestJSQuery());
-      for await (const row of result.yieldRowObjectJs()) {
-        assert.deepEqual(row, createTestJSRowObjectsJS());
-      }
-    });
-  });
-
-  test('iterate result stream rows json', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(
-        `from test_all_types()`,
-      );
-      for await (const row of result.yieldRowsJson()) {
-        assert.deepEqual(row, createTestAllTypesRowsJson());
-      }
-    });
-  });
-
-  test('iterate result stream object json', async () => {
-    await withConnection(async (connection) => {
-      const result = await connection.stream(
-        `from test_all_types()`,
-      );
-      for await (const row of result.yieldRowObjectJson()) {
-        assert.deepEqual(row, createTestAllTypesRowObjectsJson());
-      }
-    });
-  });
-
-  test('getTableNames (single, qualified)', async () => {
-    await withConnection(async (connection) => {
-      assert.deepEqual(connection.getTableNames('from memory.main.t1', true), [
-        'memory.main.t1',
-      ]);
-    });
-  });
-  test('getTableNames (single, unqualified)', async () => {
-    await withConnection(async (connection) => {
-      assert.deepEqual(connection.getTableNames('from memory.main.t1', false), [
-        't1',
-      ]);
-    });
-  });
-  test('getTableNames (multiple, qualified)', async () => {
-    await withConnection(async (connection) => {
-      assert.deepEqual(
-        connection.getTableNames(
-          'from memory.main.t1, memory.main.t2, memory.main.t1',
-          true,
-        ),
-        ['memory.main.t1', 'memory.main.t2'],
-      );
-    });
-  });
-  test('getTableNames (multiple, unqualified)', async () => {
-    await withConnection(async (connection) => {
-      assert.deepEqual(
-        connection.getTableNames(
-          'from memory.main.t1, memory.main.t2, memory.main.t1',
-          false,
-        ),
-        ['t1', 't2'],
-      );
-    });
-  });
-  test('getTableNames (none))', async () => {
-    await withConnection(async (connection) => {
-      assert.deepEqual(connection.getTableNames('select 1', true), []);
-    });
-  });
-
-  // VARIANT --------------------------------------------------------------
-  describe('VARIANT', () => {
-    test('column type and SQL NULL row', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(`select NULL::VARIANT as v`);
-        assertColumns(result, [{ name: 'v', type: VARIANT }]);
-        const chunk = await result.fetchChunk();
-        assert.isDefined(chunk);
-        if (chunk) {
-          assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-            chunk,
-            0,
-            DuckDBVariantVector,
-            [null],
-          );
-        }
-      });
-    });
-
-    test('primitive scalars round-trip through VARIANT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(`
-          select
-            true::VARIANT                              as b_true,
-            false::VARIANT                             as b_false,
-            (-128)::TINYINT::VARIANT                   as i8,
-            32767::SMALLINT::VARIANT                   as i16,
-            (-2147483648)::INTEGER::VARIANT            as i32,
-            9223372036854775807::BIGINT::VARIANT       as i64,
-            170141183460469231731687303715884105727::HUGEINT::VARIANT as i128,
-            255::UTINYINT::VARIANT                     as u8,
-            65535::USMALLINT::VARIANT                  as u16,
-            4294967295::UINTEGER::VARIANT              as u32,
-            18446744073709551615::UBIGINT::VARIANT     as u64,
-            340282366920938463463374607431768211455::UHUGEINT::VARIANT as u128,
-            (1.5)::FLOAT::VARIANT                      as f32,
-            (2.5)::DOUBLE::VARIANT                     as f64,
-            'hello'::VARCHAR::VARIANT                  as s,
-            '\\x00\\xFF'::BLOB::VARIANT                as bl,
-            (123.45)::DECIMAL(5,2)::VARIANT            as dec_s,
-            (12345678901234.5)::DECIMAL(18,1)::VARIANT as dec_64,
-            (1234567890123456789012345.6789)::DECIMAL(38,4)::VARIANT as dec_128,
-            DATE '2024-01-02'::VARIANT                 as d,
-            TIME '12:34:56.789'::VARIANT               as t_us,
-            TIMESTAMP '2024-01-02 03:04:05.678'::VARIANT as ts_us,
-            TIMESTAMP_S '2024-01-02 03:04:05'::VARIANT  as ts_s,
-            TIMESTAMP_MS '2024-01-02 03:04:05.678'::VARIANT as ts_ms,
-            TIMESTAMP_NS '2024-01-02 03:04:05.678901234'::VARIANT as ts_ns,
-            INTERVAL '3 months 4 days 5 minutes'::VARIANT as iv,
-            '00112233-4455-6677-8899-aabbccddeeff'::UUID::VARIANT as uu
-        `);
-        assertColumns(result, [
-          { name: 'b_true', type: VARIANT },
-          { name: 'b_false', type: VARIANT },
-          { name: 'i8', type: VARIANT },
-          { name: 'i16', type: VARIANT },
-          { name: 'i32', type: VARIANT },
-          { name: 'i64', type: VARIANT },
-          { name: 'i128', type: VARIANT },
-          { name: 'u8', type: VARIANT },
-          { name: 'u16', type: VARIANT },
-          { name: 'u32', type: VARIANT },
-          { name: 'u64', type: VARIANT },
-          { name: 'u128', type: VARIANT },
-          { name: 'f32', type: VARIANT },
-          { name: 'f64', type: VARIANT },
-          { name: 's', type: VARIANT },
-          { name: 'bl', type: VARIANT },
-          { name: 'dec_s', type: VARIANT },
-          { name: 'dec_64', type: VARIANT },
-          { name: 'dec_128', type: VARIANT },
-          { name: 'd', type: VARIANT },
-          { name: 't_us', type: VARIANT },
-          { name: 'ts_us', type: VARIANT },
-          { name: 'ts_s', type: VARIANT },
-          { name: 'ts_ms', type: VARIANT },
-          { name: 'ts_ns', type: VARIANT },
-          { name: 'iv', type: VARIANT },
-          { name: 'uu', type: VARIANT },
-        ]);
-        const chunk = await result.fetchChunk();
-        assert.isDefined(chunk);
-        if (!chunk) return;
-
-        // 2024-01-02 = 1970-01-01 + 19724 days; ts is 1704164645 unix seconds.
-        let i = 0;
-        const expect = (v: DuckDBValue, t: DuckDBType) =>
-          assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-            chunk,
-            i++,
-            DuckDBVariantVector,
-            [variantValue(v, t)],
-          );
-
-        expect(true, BOOLEAN);
-        expect(false, BOOLEAN);
-        expect(-128, TINYINT);
-        expect(32767, SMALLINT);
-        expect(-2147483648, INTEGER);
-        expect(9223372036854775807n, BIGINT);
-        expect(170141183460469231731687303715884105727n, HUGEINT);
-        expect(255, UTINYINT);
-        expect(65535, USMALLINT);
-        expect(4294967295, UINTEGER);
-        expect(18446744073709551615n, UBIGINT);
-        expect(340282366920938463463374607431768211455n, UHUGEINT);
-        expect(1.5, FLOAT);
-        expect(2.5, DOUBLE);
-        expect('hello', VARCHAR);
-        expect(blobValue(new Uint8Array([0x00, 0xff])), BLOB);
-        expect(decimalValue(12345n, 5, 2), DECIMAL(5, 2));
-        expect(decimalValue(123456789012345n, 18, 1), DECIMAL(18, 1));
-        expect(decimalValue(12345678901234567890123456789n, 38, 4), DECIMAL(38, 4));
-        expect(dateValue(19724), DATE);
-        expect(
-          timeValue(BigInt(((12 * 60 + 34) * 60 + 56) * 1_000_000 + 789_000)),
-          TIME,
-        );
-        expect(timestampValue(1704164645678000n), TIMESTAMP);
-        expect(timestampSecondsValue(1704164645n), TIMESTAMP_S);
-        expect(timestampMillisValue(1704164645678n), TIMESTAMP_MS);
-        expect(timestampNanosValue(1704164645678901234n), TIMESTAMP_NS);
-        expect(intervalValue(3, 4, BigInt(5 * 60 * 1_000_000)), INTERVAL);
-        expect(uuidValue(0x00112233_4455_6677_8899_aabbccddeeffn), UUID);
-      });
-    });
-
-    test('OBJECT decodes to DuckDBStructValue', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select '{"a": 1, "b": "hi", "c": null}'::JSON::VARIANT as v`,
-        );
-        assertColumns(result, [{ name: 'v', type: VARIANT }]);
-        const chunk = await result.fetchChunk();
-        // JSON integers decode as UINT64 (variant tag); 1 becomes 1n with
-        // UBIGINT type. The decoder reconstructs STRUCT field types from
-        // each child's tag.
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [
-            variantValue(
-              structValue({ a: 1n, b: 'hi', c: null }),
-              STRUCT({ a: UBIGINT, b: VARCHAR, c: SQLNULL }),
-            ),
-          ],
-        );
-      });
-    });
-
-    test('ARRAY decodes to DuckDBListValue with heterogeneous items', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select '[1, "two", null, true]'::JSON::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        // Non-null children carry differing tags so the list stays
-        // LIST(VARIANT) with each non-null item wrapped. A null child
-        // remains bare to match how a column-level LIST(VARIANT) decodes
-        // a NULL element.
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [
-            variantValue(
-              listValue([
-                variantValue(1n, UBIGINT),
-                variantValue('two', VARCHAR),
-                null,
-                variantValue(true, BOOLEAN),
-              ]),
-              LIST(VARIANT),
-            ),
-          ],
-        );
-      });
-    });
-
-    test('nested OBJECT containing ARRAY exercises recursive walk', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select '{"outer": [1, {"inner": "leaf"}, [null, true]]}'::JSON::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        // The innermost [null, true] homogenizes to LIST(BOOLEAN) — null
-        // is compatible with any sibling type. The outer array remains
-        // heterogeneous (number / struct / list) so it's LIST(VARIANT).
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [
-            variantValue(
-              structValue({
-                outer: listValue([
-                  variantValue(1n, UBIGINT),
-                  variantValue(
-                    structValue({ inner: 'leaf' }),
-                    STRUCT({ inner: VARCHAR }),
-                  ),
-                  variantValue(listValue([null, true]), LIST(BOOLEAN)),
-                ]),
-              }),
-              STRUCT({ outer: LIST(VARIANT) }),
-            ),
-          ],
-        );
-      });
-    });
-
-    test('VARIANT nested inside a LIST', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(`
-          select [1::INTEGER::VARIANT, 'x'::VARCHAR::VARIANT, NULL::VARIANT] as items
-        `);
-        assertColumns(result, [
-          { name: 'items', type: LIST(VARIANT) },
-        ]);
-        const chunk = await result.fetchChunk();
-        assertValues<DuckDBListValue, DuckDBListVector>(
-          chunk!,
-          0,
-          DuckDBListVector,
-          [
-            listValue([
-              variantValue(1, INTEGER),
-              variantValue('x', VARCHAR),
-              null,
-            ]),
-          ],
-        );
-      });
-    });
-
-    test('VARIANT nested inside a STRUCT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(`
-          select {x: 42::INTEGER::VARIANT, y: 'hello'::VARCHAR::VARIANT}
-                 ::STRUCT(x VARIANT, y VARIANT) as s
-        `);
-        assertColumns(result, [
-          { name: 's', type: STRUCT({ x: VARIANT, y: VARIANT }) },
-        ]);
-        const chunk = await result.fetchChunk();
-        assertValues<DuckDBStructValue, DuckDBStructVector>(
-          chunk!,
-          0,
-          DuckDBStructVector,
-          [
-            structValue({
-              x: variantValue(42, INTEGER),
-              y: variantValue('hello', VARCHAR),
-            }),
-          ],
-        );
-      });
-    });
-
-    test('JSON converter recurses through VARIANT (heterogeneous list)', async () => {
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(`
-          select '{"a": 1, "b": [true, null, "x"], "c": 3.14}'::JSON::VARIANT as v
-        `);
-        assert.deepStrictEqual(reader.getRowObjectsJson(), [
-          {
-            v: {
-              a: '1', // JSON converter renders BIGINTs as strings
-              b: [true, null, 'x'],
-              c: 3.14,
-            },
-          },
-        ]);
-      });
-    });
-
-    test('JS converter unwraps VARIANT to plain JS shape', async () => {
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(`
-          select '{"a": 1, "b": [true, null, "x"]}'::JSON::VARIANT as v
-        `);
-        assert.deepStrictEqual(reader.getRowObjectsJS(), [
-          { v: { a: 1n, b: [true, null, 'x'] } },
-        ]);
-      });
-    });
-
-    test('JSON converter handles VARIANT nested inside a LIST', async () => {
-      // Confirms the existing LIST converter dispatches per-element to the
-      // VARIANT converter without needing variant-specific changes.
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(`
-          select [1::INTEGER::VARIANT, 'x'::VARCHAR::VARIANT, NULL::VARIANT] as items
-        `);
-        assert.deepStrictEqual(reader.getRowObjectsJson(), [
-          { items: [1, 'x', null] },
-        ]);
-      });
-    });
-
-    test('JSON converter handles VARIANT nested inside a STRUCT', async () => {
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(`
-          select {x: 42::INTEGER::VARIANT, y: 'hello'::VARCHAR::VARIANT}
-                 ::STRUCT(x VARIANT, y VARIANT) as s
-        `);
-        assert.deepStrictEqual(reader.getRowObjectsJson(), [
-          { s: { x: 42, y: 'hello' } },
-        ]);
-      });
-    });
-
-    test('VARIANT-of-MAP unwraps via the variant converter', () => {
-      // VARIANT decoding never produces MAP values, but if a caller wraps
-      // a DuckDBMapValue in a DuckDBVariantValue (e.g. for testing or via
-      // a future write path) the converter should still produce the same
-      // {key, value}[] shape that objectArrayFromMapValue produces.
-      const v = variantValue(
-        mapValue([
-          { key: 'a', value: 1 },
-          { key: 'b', value: 2n },
-        ]),
-      );
-      assert.deepStrictEqual(JSDuckDBValueConverter(v, VARIANT, JSDuckDBValueConverter), [
-        { key: 'a', value: 1 },
-        { key: 'b', value: 2n },
-      ]);
-    });
-
-    test('VARIANT-of-UNION unwraps via the variant converter', () => {
-      const v = variantValue(unionValue('s', 'hi'));
-      assert.deepStrictEqual(JSDuckDBValueConverter(v, VARIANT, JSDuckDBValueConverter), {
-        tag: 's',
-        value: 'hi',
-      });
-    });
-
-    test('binding a VARIANT parameter (primitive)', async () => {
-      // Replaces the previous "not yet supported" test now that createValue
-      // routes the cast through DuckDB. Both wrapped and unwrapped inputs
-      // round-trip: the wrapper carries an exact `.type`; the unwrapped
-      // value falls back to `typeForValue` inference.
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(
-          'select ?::VARIANT as v, ?::VARIANT as u',
-        );
-        prepared.bind([variantValue(42, INTEGER), 'hello'], [VARIANT, VARIANT]);
-        const reader = await prepared.runAndReadAll();
-        assert.deepStrictEqual(reader.getRowObjects(), [
-          {
-            v: variantValue(42, INTEGER),
-            // INTEGER fits in JS number; typeForValue infers INTEGER for
-            // small numbers, but 'hello' is a string → VARCHAR.
-            u: variantValue('hello', VARCHAR),
-          },
-        ]);
-      });
-    });
-
-    // Extended primitive tags (BIT, BIGNUM, TIME_TZ, GEOMETRY): these share
-    // length-prefixed encoding (BIT/BIGNUM/GEOMETRY) or use signed/unsigned
-    // variants of integer reads (TIME_TZ), and weren't exercised by the main
-    // scalar round-trip test.
-
-    test('BITSTRING round-trips through VARIANT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select '0101'::BIT::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [variantValue(bitValue('0101'), BIT)],
-        );
-      });
-    });
-
-    test('BIGNUM round-trips through VARIANT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select '1234567890123456789012345'::BIGNUM::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [variantValue(1234567890123456789012345n, BIGNUM)],
-        );
-      });
-    });
-
-    test('TIME WITH TIME ZONE round-trips through VARIANT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select TIME WITH TIME ZONE '12:34:56.789+05:30'::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [
-            variantValue(
-              timeTZValue(45296789000n, 5 * 3600 + 30 * 60),
-              TIMETZ,
-            ),
-          ],
-        );
-      });
-    });
-
-    test('GEOMETRY round-trips through VARIANT', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select 'POINT(1 2)'::GEOMETRY::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        // WKB for POINT(1, 2), little-endian: 1 | 01 00 00 00 | 8-byte LE
-        // double 1.0 | 8-byte LE double 2.0.
-        const pointWkb = new Uint8Array([
-          0x01, 0x01, 0x00, 0x00, 0x00,
-          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xf0, 0x3f,
-          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40,
-        ]);
-        assertValues<DuckDBVariantValue, DuckDBVariantVector>(
-          chunk!,
-          0,
-          DuckDBVariantVector,
-          [variantValue(geometryValue(pointWkb), GEOMETRY)],
-        );
-      });
-    });
-
-    test('DuckDBGeometryType exposes the CRS from the logical type', async () => {
-      // Direct construction: crs is preserved, and the cached `instance` is
-      // only used when both alias and crs are absent.
-      const explicit = DuckDBGeometryType.create(undefined, 'GEOGCRS["x"]');
-      assert.strictEqual(explicit.crs, 'GEOGCRS["x"]');
-      assert.notStrictEqual(explicit, DuckDBGeometryType.instance);
-      assert.strictEqual(DuckDBGeometryType.create(), DuckDBGeometryType.instance);
-      assert.strictEqual(DuckDBGeometryType.instance.crs, undefined);
-
-      // toString / toJson include the CRS when present.
-      assert.strictEqual(DuckDBGeometryType.instance.toString(), 'GEOMETRY');
-      assert.strictEqual(explicit.toString(), `GEOMETRY('GEOGCRS["x"]')`);
-      assert.deepStrictEqual(DuckDBGeometryType.instance.toJson(), {
-        typeId: DuckDBTypeId.GEOMETRY,
-      });
-      assert.deepStrictEqual(explicit.toJson(), {
-        typeId: DuckDBTypeId.GEOMETRY,
-        crs: 'GEOGCRS["x"]',
-      });
-      // CRS strings containing single quotes are escaped in toString.
-      assert.strictEqual(
-        new DuckDBGeometryType(undefined, `it's`).toString(),
-        `GEOMETRY('it''s')`,
-      );
-
-      // Round-trip through SQL: `column_logical_type` -> `asType()` should
-      // produce a DuckDBGeometryType carrying the CRS.
-      await withConnection(async (connection) => {
-        const withCrs = await connection.run(
-          `SELECT 'POINT(1 2)'::GEOMETRY('GEOGCRS["x"]') AS g`,
-        );
-        const withCrsType = withCrs.columnTypes()[0];
-        assert.ok(withCrsType instanceof DuckDBGeometryType);
-        assert.strictEqual(withCrsType.crs, 'GEOGCRS["x"]');
-
-        const plain = await connection.run(`SELECT 'POINT(1 2)'::GEOMETRY AS g`);
-        const plainType = plain.columnTypes()[0];
-        assert.ok(plainType instanceof DuckDBGeometryType);
-        assert.strictEqual(plainType.crs, undefined);
-      });
-    });
-
-    // Regression tests for code-review fixes.
-
-    test('UINTEGER max routes through the JS / JSON converters', async () => {
-      // The decoder records the exact UINTEGER type on the wrapper, so the
-      // converter dispatches via UINTEGER → numberFromValue and the value
-      // stays a JS number. (Without the type hint, `typeForValue` would
-      // infer BIGINT for a number > INT32 max and the variant walker
-      // would coerce to bigint as a fallback — exercised separately for
-      // unwrapped numeric values.)
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(
-          `select 4294967295::UINTEGER::VARIANT as u`,
-        );
-        assert.deepStrictEqual(reader.getRowObjectsJS(), [
-          { u: 4294967295 },
-        ]);
-        assert.deepStrictEqual(reader.getRowObjectsJson(), [
-          { u: 4294967295 },
-        ]);
-      });
-    });
-
-    test('setItem(null) and flush() on a VARIANT vector are no-ops', async () => {
-      await withConnection(async (connection) => {
-        const result = await connection.run(
-          `select 42::INTEGER::VARIANT as v`,
-        );
-        const chunk = await result.fetchChunk();
-        const vec = chunk!.getColumnVector(0) as DuckDBVariantVector;
-        // No throw: parent containers (struct/list/array) call setItem(null)
-        // and flush() generically on every child; VARIANT must tolerate
-        // those even though writes aren't supported.
-        vec.setItem(0, null);
-        vec.flush();
-        // Writing a non-null value remains an error.
-        assert.throws(
-          () => vec.setItem(0, variantValue(1)),
-          /Setting VARIANT values is not yet supported/,
-        );
-      });
-    });
-
-    test('STRUCT(VARIANT, INTEGER) can be read end-to-end', async () => {
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(`
-          select {x: 42::INTEGER::VARIANT, y: 7::INTEGER}
-                 ::STRUCT(x VARIANT, y INTEGER) as s
-        `);
-        assert.deepStrictEqual(reader.getRowObjectsJS(), [
-          { s: { x: 42, y: 7 } },
-        ]);
-      });
-    });
-
-    // Binding / appending VARIANTs. DuckDB has no `duckdb_create_variant`
-    // primitive; instead the variant target type triggers an implicit cast
-    // from whatever Value we hand it. Four routes get exercised below:
-    //   1. Wrapper round-trip via `bindVariant` / `appendVariant`.
-    //   2. Generic `bindValue` / `appendValue` with target=VARIANT.
-    //   3. Generic `bindValue` / `appendValue` with a *non-VARIANT* target
-    //      (DuckDB performs the cast on its side).
-    //   4. Type-specific append methods (`appendInteger`, `appendVarchar`)
-    //      that bypass `createValue` entirely.
-
-    test('bindVariant round-trips a primitive VARIANT (path 1)', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select 42::INTEGER::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        prepared.bindVariant(1, decoded);
-        const r2 = await prepared.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects(), [
-          { v: variantValue(42, INTEGER) },
-        ]);
-      });
-    });
-
-    test('bindVariant round-trips a STRUCT VARIANT (path 1)', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '{"a": 1, "b": "hi"}'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        prepared.bindVariant(1, decoded);
-        const r2 = await prepared.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects(), [
-          {
-            v: variantValue(
-              structValue({ a: 1n, b: 'hi' }),
-              STRUCT({ a: UBIGINT, b: VARCHAR }),
-            ),
-          },
-        ]);
-      });
-    });
-
-    test('bindVariant round-trips a homogeneous ARRAY VARIANT (path 1)', async () => {
-      // A JSON array whose elements all share a tag becomes
-      // LIST(commonType) with bare items, which round-trips cleanly.
-      // (Heterogeneous arrays decode as LIST(VARIANT) with wrapped items
-      // and currently cannot be re-bound — see the limitation note.)
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '[1, 2, 3]'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        prepared.bindVariant(1, decoded);
-        const r2 = await prepared.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects(), [
-          {
-            v: variantValue(
-              listValue([1n, 2n, 3n]),
-              LIST(UBIGINT),
-            ),
-          },
-        ]);
-      });
-    });
-
-    test('bindValue with explicit VARIANT target (path 2)', async () => {
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(
-          `select ?::VARIANT as v, ?::VARIANT as u`,
-        );
-        // Mixed: a raw number (typeForValue → INTEGER) and a raw string
-        // (typeForValue → VARCHAR).
-        prepared.bindValue(1, 42, VARIANT);
-        prepared.bindValue(2, 'hello', VARIANT);
-        const reader = await prepared.runAndReadAll();
-        assert.deepStrictEqual(reader.getRowObjects(), [
-          {
-            v: variantValue(42, INTEGER),
-            u: variantValue('hello', VARCHAR),
-          },
-        ]);
-      });
-    });
-
-    test('bindValue with explicit VARIANT target accepts null (path 2)', async () => {
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        prepared.bindValue(1, null, VARIANT);
-        const reader = await prepared.runAndReadAll();
-        assert.deepStrictEqual(reader.getRowObjects(), [{ v: null }]);
-      });
-    });
-
-    test('bindValue with non-VARIANT target casts on DuckDB side (path 3)', async () => {
-      // The target column is VARIANT (via the `?::VARIANT` cast) but the
-      // bound Value is INTEGER / VARCHAR / STRUCT. DuckDB performs the
-      // cast.
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(
-          `select ?::VARIANT as i, ?::VARIANT as s, ?::VARIANT as o`,
-        );
-        prepared.bindValue(1, 42, INTEGER);
-        prepared.bindValue(2, 'hello', VARCHAR);
-        prepared.bindValue(
-          3,
-          structValue({ a: 1, b: 'x' }),
-          STRUCT({ a: INTEGER, b: VARCHAR }),
-        );
-        const reader = await prepared.runAndReadAll();
-        assert.deepStrictEqual(reader.getRowObjects(), [
-          {
-            i: variantValue(42, INTEGER),
-            s: variantValue('hello', VARCHAR),
-            o: variantValue(
-              structValue({ a: 1, b: 'x' }),
-              STRUCT({ a: INTEGER, b: VARCHAR }),
-            ),
-          },
-        ]);
-      });
-    });
-
-    test('appendVariant + appendValue + type-specific append (paths 1, 2, 3, 4)', async () => {
-      await withConnection(async (connection) => {
-        await connection.run(`CREATE TABLE t (v VARIANT)`);
-        const appender = await connection.createAppender('t');
-        // Path 1: appendVariant with a wrapped value carrying its exact type.
-        appender.appendVariant(variantValue(1, INTEGER));
-        appender.endRow();
-        // Path 2: appendValue with target=VARIANT, raw input.
-        appender.appendValue('hello', VARIANT);
-        appender.endRow();
-        // Path 3: appendValue with target=INTEGER against a VARIANT column.
-        appender.appendValue(3, INTEGER);
-        appender.endRow();
-        // Path 4: type-specific append bypasses createValue.
-        appender.appendInteger(4);
-        appender.endRow();
-        // Null.
-        appender.appendNull();
-        appender.endRow();
-        appender.closeSync();
-
-        const reader = await connection.runAndReadAll(`select v from t`);
-        assert.deepStrictEqual(reader.getRowObjects(), [
-          { v: variantValue(1, INTEGER) },
-          { v: variantValue('hello', VARCHAR) },
-          { v: variantValue(3, INTEGER) },
-          { v: variantValue(4, INTEGER) },
-          { v: null },
-        ]);
-      });
-    });
-
-    test('heterogeneous LIST(VARIANT) cannot be bound directly', async () => {
-      // The C API's create_list_value rejects a list of values with
-      // heterogeneous physical types, so a decoded `LIST(VARIANT)` (which
-      // wraps each non-null element to preserve mixed types) cannot be
-      // bound back through a prepared statement. Pin the failure so we
-      // notice if the surface changes. The error wording originates in
-      // the native bindings — if it changes phrasing across DuckDB
-      // versions, update the regex.
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '[1, "two"]'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        // Confirm the decoder produced the heterogeneous shape.
-        assert.deepStrictEqual(decoded.type, LIST(VARIANT));
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        assert.throws(
-          () => prepared.bindVariant(1, decoded),
-          /Failed to create (?:VARIANT value.*?Failed to create )?list value/,
-        );
-      });
-    });
-
-    // Round-trip coverage for the SQLNULL-compatible homogenize logic.
-
-    test('[1, 2, null] homogenizes to LIST(UBIGINT) with a bare null', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '[1, 2, null]'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        assert.deepStrictEqual(decoded, variantValue(
-          listValue([1n, 2n, null]),
-          LIST(UBIGINT),
-        ));
-        // And round-trips back through bindVariant.
-        const p = await connection.prepare(`select ?::VARIANT as v`);
-        p.bindVariant(1, decoded);
-        const r2 = await p.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects()[0].v, decoded);
-      });
-    });
-
-    test('[null, null] collapses to LIST(SQLNULL) and round-trips', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '[null, null]'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        assert.deepStrictEqual(decoded, variantValue(
-          listValue([null, null]),
-          LIST(SQLNULL),
-        ));
-        const p = await connection.prepare(`select ?::VARIANT as v`);
-        p.bindVariant(1, decoded);
-        const r2 = await p.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects()[0].v, decoded);
-      });
-    });
-
-    test('empty [] decodes as LIST(SQLNULL) and round-trips', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '[]'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        assert.deepStrictEqual(decoded, variantValue(listValue([]), LIST(SQLNULL)));
-        const p = await connection.prepare(`select ?::VARIANT as v`);
-        p.bindVariant(1, decoded);
-        const r2 = await p.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects()[0].v, decoded);
-      });
-    });
-
-    test('empty {} decodes as STRUCT() and round-trips', async () => {
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '{}'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        assert.deepStrictEqual(decoded, variantValue(structValue({}), STRUCT({})));
-        const p = await connection.prepare(`select ?::VARIANT as v`);
-        p.bindVariant(1, decoded);
-        const r2 = await p.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects()[0].v, decoded);
-      });
-    });
-
-    test('STRUCT VARIANT containing a SQLNULL-typed field round-trips', async () => {
-      // Covers the case where a struct field is JSON-null; the decoder
-      // captures the field type as SQLNULL and the bind path must accept
-      // it (top-level null shortcut in createValue handles the inner null).
-      await withConnection(async (connection) => {
-        const r1 = await connection.runAndReadAll(
-          `select '{"a": 1, "b": null}'::JSON::VARIANT as v`,
-        );
-        const decoded = r1.getRowObjects()[0].v as DuckDBVariantValue;
-        const p = await connection.prepare(`select ?::VARIANT as v`);
-        p.bindVariant(1, decoded);
-        const r2 = await p.runAndReadAll();
-        assert.deepStrictEqual(r2.getRowObjects()[0].v, decoded);
-      });
-    });
-
-    test('bindVariant accepts null (short-circuits to bindNull)', async () => {
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        prepared.bindVariant(1, null);
-        const reader = await prepared.runAndReadAll();
-        assert.deepStrictEqual(reader.getRowObjects(), [{ v: null }]);
-      });
-    });
-
-    test('mismatched DuckDBVariantValue wrapper produces a VARIANT-context error', async () => {
-      // `variantValue('hello', INTEGER)` claims the inner value is an
-      // integer but it's actually a string. The error message should
-      // surface the VARIANT context to aid debugging.
-      await withConnection(async (connection) => {
-        const prepared = await connection.prepare(`select ?::VARIANT as v`);
-        assert.throws(
-          () => prepared.bindVariant(1, variantValue('hello', INTEGER)),
-          /Failed to create VARIANT value.*input is not a number/,
-        );
-      });
-    });
-  });
-
-  // DuckDBUUIDValue.fromUint128 ----------------------------------------------
-  // Regression coverage for the MSB-flip-with-signed-storage contract:
-  // DuckDBUUIDValue always holds DuckDB's signed-int128 storage form, and
-  // `fromUint128` must convert into that form (including for inputs whose
-  // unsigned value sits below 2**127, which after the MSB flip end up with
-  // bit 127 set and therefore need a negative bigint to represent).
-  describe('DuckDBUUIDValue.fromUint128', () => {
-    test('low-MSB input maps to a negative signed-int128 hugeint', async () => {
-      // 0x0011… has bit 127 clear; after MSB flip it has bit 127 set, which
-      // must surface as a negative bigint to match DuckDB's int128 storage.
-      const v = uuidValue(0x00112233_4455_6677_8899_aabbccddeeffn);
-      assert.strictEqual(
-        v.hugeint,
-        BigInt.asIntN(128, 0x80112233_4455_6677_8899_aabbccddeeffn),
-      );
-      assert.isTrue(v.hugeint < 0n, 'hugeint should be negative');
-      assert.strictEqual(v.toString(), '00112233-4455-6677-8899-aabbccddeeff');
-      // And: the same UUID round-tripped through a UUID column matches
-      // exactly (the column decoder uses `fromStoredHugeInt` + `getInt128`).
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(
-          `select '00112233-4455-6677-8899-aabbccddeeff'::UUID as u`,
-        );
-        const decoded = reader.getRows()[0][0];
-        assert.deepStrictEqual(decoded, v);
-      });
-    });
-
-    test('high-MSB input maps to a positive hugeint', async () => {
-      // 0xf0e1… has bit 127 set; after MSB flip the high bit is clear and
-      // the value fits in a positive signed int128.
-      const v = uuidValue(0xf0e1d2c3b4a596870123456789abcdefn);
-      assert.strictEqual(
-        v.hugeint,
-        0x70e1d2c3b4a596870123456789abcdefn,
-      );
-      assert.isTrue(v.hugeint > 0n, 'hugeint should be positive');
-      assert.strictEqual(v.toString(), 'f0e1d2c3-b4a5-9687-0123-456789abcdef');
-      await withConnection(async (connection) => {
-        const reader = await connection.runAndReadAll(
-          `select 'f0e1d2c3-b4a5-9687-0123-456789abcdef'::UUID as u`,
-        );
-        const decoded = reader.getRows()[0][0];
-        assert.deepStrictEqual(decoded, v);
-      });
-    });
-
-    test('toUint128 round-trips both halves of the range', () => {
-      const low = uuidValue(0x00112233_4455_6677_8899_aabbccddeeffn);
-      assert.strictEqual(
-        low.toUint128(),
-        0x00112233_4455_6677_8899_aabbccddeeffn,
-      );
-      const high = uuidValue(0xf0e1d2c3b4a596870123456789abcdefn);
-      assert.strictEqual(
-        high.toUint128(),
-        0xf0e1d2c3b4a596870123456789abcdefn,
-      );
-      // Extremes.
-      const allZeros = uuidValue(0n);
-      assert.strictEqual(allZeros.toUint128(), 0n);
-      const allOnes = uuidValue(2n ** 128n - 1n);
-      assert.strictEqual(allOnes.toUint128(), 2n ** 128n - 1n);
-    });
-  });
-
-  // varintDecode (LEB128) ----------------------------------------------------
-  describe('varintDecode', () => {
-    function makeView(bytes: number[]): DataView {
-      const buf = new Uint8Array(bytes);
-      return new DataView(buf.buffer, buf.byteOffset, buf.byteLength);
-    }
-    test('single-byte values (< 0x80)', () => {
-      assert.deepStrictEqual(varintDecode(makeView([0x00]), 0), {
-        value: 0,
-        nextOffset: 1,
-      });
-      assert.deepStrictEqual(varintDecode(makeView([0x01]), 0), {
-        value: 1,
-        nextOffset: 1,
-      });
-      assert.deepStrictEqual(varintDecode(makeView([0x7f]), 0), {
-        value: 127,
-        nextOffset: 1,
-      });
-    });
-    test('two-byte values', () => {
-      // 128 = 0b1000_0000 → bytes [0x80, 0x01]
-      assert.deepStrictEqual(varintDecode(makeView([0x80, 0x01]), 0), {
-        value: 128,
-        nextOffset: 2,
-      });
-      // 300 = 0b1_0010_1100 → bytes [0xac, 0x02]
-      assert.deepStrictEqual(varintDecode(makeView([0xac, 0x02]), 0), {
-        value: 300,
-        nextOffset: 2,
-      });
-    });
-    test('respects starting offset', () => {
-      // Skip a leading byte; decode 300 starting at offset 1.
-      const view = makeView([0xff, 0xac, 0x02, 0x99]);
-      assert.deepStrictEqual(varintDecode(view, 1), {
-        value: 300,
-        nextOffset: 3,
-      });
-    });
-    test('five-byte value (uint32 max)', () => {
-      // 0xffffffff → bytes [0xff, 0xff, 0xff, 0xff, 0x0f]
-      assert.deepStrictEqual(
-        varintDecode(makeView([0xff, 0xff, 0xff, 0xff, 0x0f]), 0),
-        { value: 0xffffffff, nextOffset: 5 },
-      );
-    });
-    test('throws on overflow (6+ byte varint)', () => {
-      // Six continuation bytes would shift past 32 bits.
-      assert.throws(
-        () =>
-          varintDecode(
-            makeView([0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x01]),
-            0,
-          ),
-        /varint overflow/,
-      );
-    });
-    test('throws on overflow (5-byte value > uint32 max)', () => {
-      // [0xff,0xff,0xff,0xff,0x10] decodes to 2**32 — one above uint32 max.
-      assert.throws(
-        () =>
-          varintDecode(makeView([0xff, 0xff, 0xff, 0xff, 0x10]), 0),
-        /varint overflow/,
-      );
-      // [0xff,0xff,0xff,0xff,0x7f] decodes to 2**35-1 — a 35-bit value
-      // that the previous (looser) guard accepted.
-      assert.throws(
-        () =>
-          varintDecode(makeView([0xff, 0xff, 0xff, 0xff, 0x7f]), 0),
-        /varint overflow/,
-      );
-    });
-    test('throws on truncated input (continuation bit set on final byte)', () => {
-      assert.throws(
-        () => varintDecode(makeView([0x80]), 0),
-        /varint truncated/,
-      );
-      assert.throws(
-        () => varintDecode(makeView([0xff, 0xff, 0xff]), 0),
-        /varint truncated/,
-      );
-    });
-  });
-
-});
